Bug 1941438 - Prevent notification workers from triggering notification prompt r=android-reviewers,twhite
Though the bug is about not showing the prompt before ToS onboarding card, I believe that the actual issue lies in the way we handle notification workers. Two features, reengagment prompt and nimbus messeges, depend on notifications, but we launch respective workers without considering if notifications are enabled or not. If conditions are met and workers decide to show notification, it will trigger enable notification prompt during onboarding, even before the user saw the onboarding card. What is worse, it might shown on top of the setting default browser prompt.

There is a number of patches around this logic:
https://github.com/mozilla-mobile/firefox-android/commit/d9a70827c37cc5b808e358d6d38171a369a50d40 - added a check for a marketing channel before launching notificaton workers; this worked in couple with `showNotificationPermissionPromptIfRequired`, ensuring that the user enabled notifications first, before we start the workers.

My guess is there was an issue with marketing channel id, that prevented some deviced from launching the reengagment prompt, so that condition to start workers was removed in https://github.com/mozilla-mobile/firefox-android/pull/932/files. `showNotificationPermissionPromptIfRequired` used to set the marketing channel, but it was removed in https://github.com/mozilla-mobile/firefox-android/commit/d10db8751f36ee4c8ad4d7108daacb3aa436ce7f as well.

Now, my gut feeling is that we do not want and should not start notification workers unless the user has enabled notifications. It's a different condition from the marketing channel, and it sounds like a logical prerequisite to start the two features.

Another approach we could take would be introducing a check if we are about to show onboarding and if the onboarding has a ToS card, and after that card has been shown, set up workers. In that case, though, we also end up with notification prompt showing on top of the default browser prompt before user has seen notification onboarding card. Hence, I went with this approach.

Differential Revision: https://phabricator.services.mozilla.com/D235101

UltraBlame original commit: bb4b7b46eeb663ff8b7cdf40a8d3d3d9aa0c2bd5