Bug 1924066 - part5 : make the test more robust. r=media-playback-reviewers,padenot
As we're just caring what decoder is being used, we can set video as loop
which prevents video decoder from being destroyed when the decoding is
completed.

In addition, only waiting for the play promise is enough, because it will
be resolved when the first frame is decoded, which means the decoder has
been determined.

Depends on D234761

Differential Revision: https://phabricator.services.mozilla.com/D234762

UltraBlame original commit: 2ab38f908263fc74938c293fca89c0336c78aedc