Bug 1942658 - Move the extraParams from the Nimbus search section to the search configuration section. r=search-reviewers,scunnane
This will mean that the search section will be exclusively for features, and the configuration section will be items relating more to the configuration.
Hence this should provide a better split for experiments to be able to use.

Differential Revision: https://phabricator.services.mozilla.com/D234888

UltraBlame original commit: 91d5a0bcc10e799dbfe7f0cab59b84d00cb008e5