Bug 1937801 - [ci] Enable pip and uv caches for test tasks, r=taskgraph-reviewers,jmaher
The inclusion of 'checkout' might be surprising given the last commit is
removing the 'checkout' cache. But Taskgraph's default logic is to only
enable the checkout cache if a checkout is actually being used. So it's
safe to include here in case a test task *does* use a checkout.

Differential Revision: https://phabricator.services.mozilla.com/D235627

UltraBlame original commit: f35e913efae9381517562d507c79b3ce75e11dc4