Bug 1943345 [wpt PR 50243] - Adjust H264 profile-level-id test to avoid false failures., a=testonly
Automatic update from web-platform-tests
Adjust H264 profile-level-id test to avoid false failures.

In preparation of fixing the bug that this test gives coverage for, it
needs to be updated in order to avoid it failing even when things are
working as expected.

If the codec is TRULY recvonly, then it is correct to only negotiate
VP8 and the assertion that H264 needs to be negotiated is wrong.

What we need is a profile-level-id that is "recvonly enough" to only be
listed in RTCRtpReceiver.getCapabilities() but for which we are actually
capable of sending, i.e. there exists some other profile-level-id in
RTCRtpSender.getCapabilities() that is not "==" equal but that is
equal if you "ignoreLevels = true".

64001f is one such profile that is commonly available (level 3.1). On
MacBook M1 Pro where we can send 5.2, this test is applicable.

This test will pass once this WebRTC CL lands:
https://webrtc-review.googlesource.com/c/src/+/374043

Bug: chromium:381407888
Change-Id: I4010d76092bf9b51b20f93e756e2dacf3e8665a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6194115
Reviewed-by: Harald Alvestrand <htachromium.org>
Commit-Queue: Henrik Boström <hboschromium.org>
Cr-Commit-Position: refs/heads/main{#1410504}

--

wpt-commits: 33c6f84909c086390c57bdb5cf838f2ff5bf1729
wpt-pr: 50243

UltraBlame original commit: ef5e95a648916401bc7fa9bcfc09fe2cdefb814c