Bug 1901467: Implement ITextRangeProvider::FindAttribute, r=Jamie
This revision implements UIA function ITextRangeProvider::FindAttribute,
which allows clients to search within a UiaTextRange for a range where a
given attribute matches a given value. This revision adds the
functionality and tests for that functionality. On the way, this
revision adds GetAttribute functionality for TextLeafPoint (it already
existed for TextLeafRange). This is useful for checking single points
without wasting work constructing or searching a range.

Differential Revision: https://phabricator.services.mozilla.com/D235395

UltraBlame original commit: c4818867fa0121bf529d57292d97cd0bd1c8012e