Bug 1944995: Use MinTimeT as the start value for clamping time values. r=spidermonkey-reviewers,dminor
`MinTimeT` is either `StartOfTime / msPerSecond` (when `JS_HAS_INTL_API`) and
otherwise 0. Clamping to a non-negative value is needed for `localtime`, in which
case `MinTimeT` is 0. When instead using ICU, we don't actually need to clamp, but
because other code in this file asserts only clamped values are used, we have to
clamp this to `MinTimeT + SecondsPerDay`, because clamping to only `SecondsPerDay`
means using Jan 2 1970 as the start value.

Differential Revision: https://phabricator.services.mozilla.com/D236534

UltraBlame original commit: 00aefe73c20c4e29fcee910be336e9b141a2ed64