Bug 1944797 - Part 1: Add LIRGeneratorShared::useLowWord to allocate register for int64 low-word. r=jandem
`LIRGeneratorShared::use` allowed to allocate a register for the low-word of an
int64 value, because the `INT64_PIECES > 1` assertion didn't propery work. (The
assertion was wrapped with `#if INT64_PIECES > 1`, but `INT64_PIECES` is a C++
constant and not a macro processor `#define`, so `#if INT64_PIECES > 1` always
evaluated to `false`.)

Add `LIRGeneratorShared::useLowWord` for 32-bit targets to provide the correct
way to allocate a register for an int64 low-word. And then update the three places
which only worked due to the incorrect assertion:

- `lowerForShiftInt64` on x86 and arm32.
- `visitWasmWrapU32Index` when `JS_64BIT` is false.
- `visitWasmStore` for x86 when the input is an int64 value.

Differential Revision: https://phabricator.services.mozilla.com/D236537

UltraBlame original commit: 12ba7ff7a874746f6bae71f3e7c77881bfde86ca