Bug 1945290 - Part 1: Hide `TelemetryReportingPolicyImpl`. r=chutten,mviar
This partially reverts the changes of an earlier commit, which made
some things asynchronous.  The `NimbusFeatures.getAllVariables()` API
is not, in fact, asynchronous, so opening the internals is not
strictly necessary.  This simplifies the expression and testing for
later.

N.b.: the existing implementation did not correctly test
`_openFirstRunPage`: that function was async and there was no `await`.
Therefore existing tests were asserting things about the function
definitions and not the function behaviours.

Differential Revision: https://phabricator.services.mozilla.com/D235886

UltraBlame original commit: caf8581c6983dfb4646bddb3c68d0e07a52b36f7