Bug 1779688 - Add copy text fragment context menu items. r=mconley,fluent-reviewers,bolsson
This patch adds context menu entries to create a link to the current selection
of a site ("Text Fragment") and copies it to the clipboard.
Furthermore, if any Text Fragments are visible on the page,
it adds a context menu entry to remove these highlights.

This patch was mostly written by :jabramson.

Differential Revision: https://phabricator.services.mozilla.com/D235320

UltraBlame original commit: 21c27f2ca5fa5c934ea8ea06cb5051e3fa7ad203