Bug 1943696 - wasm: Add JitDump support for optimizing code. r=iain
This was a little trickier than I hoped due to the fact that we compile
on background threads in parallel, so we don't know the final code
address for a while. The easiest solution I found was to just hold on
to the perf spewer until the final link, then discard it.

Adds JitDump support to wasm using ion.
  - Mark offsets in CodeGenerator::generateWasm
  - Transfer the IonPerfSpewer out of CodeGenerator, collect them all in ModuleGenerator
    in preparation for the final linking of code.
  - Use the perf spewers instead of our basic perf integration when we have them.
  - Refactor the CodeBlock::sendToProfiler method.

Differential Revision: https://phabricator.services.mozilla.com/D237153

UltraBlame original commit: fbb992cdb35f5c71c60a961e8f2f85e6351c4ef8