Bug 1766133 - Make the gather profile thread joinable and join it during shutdown to avoid leaks r=aabh,profiler-reviewers DONTBUILD
Previously the gather profile thread was unjoinable. It was mostly a fire-off
thread where we create the thread from the `ProfilerChild` thread and don't
really wait for it to join. It was sending the profile back to the ProfilerChild
so we were notified that it was done. It worked fairly well in the past because
it was a short running thread that's only needed during profile capture for a
short duration.

But it was creating some leaks during shutdown when we try to capture a profile
while a process is shutting down. Let's assume that an iframe or a tab is
closing down and destroying the content process. In the meantime when we try to
capture a profile, it's still sending the GatherProfile IPC, but shutting down
before the gather profile thread could exit. That's why we were leaking
everything inside that thread.

So instead, when we are on shutdown and arrive to `ProfilerChild::ActorDestroy`,
we look at the capture profile thread handle, and check if there is a thread,
and wait for it to be joined while shutting down. This is very unlikely to hit
in real life, but it's possible to hit in CI.

Differential Revision: https://phabricator.services.mozilla.com/D237557

UltraBlame original commit: 7464dfed8db1c0e0796722c6ff64682b2415b933