Bug 1947303 - Skip dispatching AsyncSendShutDownMessage if already shut(ting) down, r=ipc-reviewers,mccr8
After the changes in bug 1942128, it is now possible for the final
KeepAlive for a process to be destroyed very late during shutdown,
during final cycle-collection. In this case, the main thread's event
target would already be dead, and the call to dispatch in
`AsyncSendShutDownMessage` crashes with a null pointer.

This adds some basic checks to make sure the `ShutDownProcess` function
will have an effect before trying to dispatch. This should ensure that
the dispatch is skipped when the actor is already dead, which will be
the case late during shutdown.

Differential Revision: https://phabricator.services.mozilla.com/D237778

UltraBlame original commit: 50128bea153cd14610b21ae1bb9c53124220da0c