Bug 1944579 - Use img instead of iframe for the Android error page's images. r=android-reviewers,ohall
When SHIP is enabled, the source assignment to the iframes holding the error page
images inserts a new history element in the tree. This causes issues in the forward
and backward navigation in the history. In order to avoid that, iframes in the error
pages are replaced with <img> elements to ensure that there's no history item added
to the tree at the time of src assignment.

Differential Revision: https://phabricator.services.mozilla.com/D237876

UltraBlame original commit: f9674f379c5e4690721966e3410de7f548c71730