Bug 1946901 - Fix race in about:translations reacting to shortlist pref r=translations-reviewers,gregtatum
Fixes a race condition where the about:translations page may attempt
to rebuild the translator before the active TranslationsEngine is fully
destroyed when the useLexicalShortlist pref is flipped. This is likely the
cause of the intermittent test failure associated with this bug.

Differential Revision: https://phabricator.services.mozilla.com/D237846

UltraBlame original commit: 148b9bdada0a6fab094569f6a910ddf32645582a