Bug 1947240 - [devtools] Fix deleting cookies with Domain attributes in Storage panel. r=devtools-reviewers,jdescottes.
The _removeCookies method had its own (faulty) host matching function,
which was buggy for cookies with domain.
The class already has a isCookieAtHost method, which doesn't have
the faulty check, so we use this instead.
A test case is added in browser_storage_cookies_delete_all.js.

Adding a new tab in the test revealed that the cleanup function was only
calling window.clear for the last opened tab, so we now do it for all the
tabs we're closing.

Differential Revision: https://phabricator.services.mozilla.com/D238012

UltraBlame original commit: 8b385a2d7056ab261c649c4263b2b34bbe188cdb