Bug 1949238: Part 1 - Expand on hover message bar in sidebar customize r=sidebar-reviewers,desktop-theme-reviewers,reusable-components-reviewers,nsharpley
This is an initial patch with the message bar styled as we would like,
but it is using temporary strings directly in the markup. These will
need to be extracted to fluent strings in a follow up change.

Differential Revision: https://phabricator.services.mozilla.com/D239026

UltraBlame original commit: c324126d242b9c16504f1386c2f82dcebd75c0ef