Bug 1950615 - show "unable to connect" Content Analysis dialog in all cases r=dlp-reviewers,handyman
I verified that this fixes the problem in this bug and does not regress
bug 1928064.

The problem was that if only one request was being made with the same
user action ID, the otherEntries would always be empty, and so we would
not show the dialog. Now we handle multiple requests by removing their
info from requestTokenToRequestInfo, and thus when we get the
"dlp-request-made" notification we will exit early and not show a dialog.

Differential Revision: https://phabricator.services.mozilla.com/D239755

UltraBlame original commit: cfc7440fb334a6a7d222e11a2818623d0bf5a627