Bug 1947431 - Implement ZstdDecompressionAlgorithms r=saschanaz,jesup
This patch introduces a DecompressionStreamAlgorithms abstract base
class, which is implemented by ZLibDecompressionStreamAlgorithms and
ZstdDecompressionStreamAlgorithms.

It also introduces a new static pref for the "zstd" format to be used
with CompressionStream and DecompressionStream, and introduces new WPTs
for zstd DecompressionStream with equivalent coverage to those for ZLib.

Differential Revision: https://phabricator.services.mozilla.com/D238685

UltraBlame original commit: 1dcbc510621be17f1ee0c35de86234cd9d334330