Bug 1665474 - Part 8: Update document.domain warning wording, r=farre,flod
Previously this warning could only fire in a cross-origin-isolated
context, but after the changes in this patch stack, it can also happen
if the webpage opts in with Origin-Agent-Cluster: ?1.

This updates the wording of the warning to more accurately reflect the
new situations which cause it to fire.

Differential Revision: https://phabricator.services.mozilla.com/D236977

UltraBlame original commit: 5a4d6c2e0dec37eb67415b43c5323d916df4bf3c