Bug 1665474 - Part 9: Handle calls to browser.createAboutBlankDocumentViewer with http(s) principals, r=farre
It is possible for frontend JS to call this method on a browser element
with an http(s) principal. In this case, treat the new about:blank
document like it was created without a Origin-Agent-Cluster header, and
ensure the map is updated before the message to create the viewer is
sent into the content process.

Differential Revision: https://phabricator.services.mozilla.com/D238115

UltraBlame original commit: 81c6cd7f8b5bd2750c00cc092d67199092b4ed87