Bug 1949091 - Do not check redirect chain when deciding if in HTTPS-Only Meta/JS redirect loop r=manuel
The assumption

> This information is only when the redirect chain is empty. When the redirect
> chain is not empty, this load is definitely triggered by redirection, not a
> user gesture.

seems to be wrong. We can land in a case where a user first triggers a
navigation themselves, and then a server-side redirect is happening. In that
case, the load was triggered by a user gesture, but the redirect chain is still
not empty.

Differential Revision: https://phabricator.services.mozilla.com/D239756

UltraBlame original commit: 96028304ba8cc101df24160e34be26bf26ae2e1c