Bug 1951362 - Text Fragments: Correctly consider prefixes for range-based matching. r=dom-core,farre
For range-based matching, the previous matches were determined incorrectly:
The `start` match must be matched until the end of the `start` of the target range,
and the `end` match must be matched until the end of the `end` of the target rage.

Before this change, the `start` of the target range was always considered as a
match to eliminate, therefore producing wrong results.

Differential Revision: https://phabricator.services.mozilla.com/D240131

UltraBlame original commit: aa91bcae0e424f187e09019307e5c205ba9e0837