Bug 1953005 - Change strategy for chrome URI registration for newtab. r=home-newtab-reviewers,extension-reviewers,robwu,thecount
In order to make activity-stream.css accessible to the newtab page, we need to register
it at a chrome URI. We were originally doing this via ComponentManager.addBootstrappedManifestLocation,
which is a holdover from old bootstrapped addons. It turns out that this only works when
the path passed to it is either to an XPI or a "real" directory on the file system. When
packaging newtab as a built-in addon, when not running an out-of-band update, it's hosted
in the omni.ja JAR file, and that's not supported by addBootstrappedManifestLocation.

This patch does two things:

1. It gets rid of the chrome.manifest file, and
2. Has resource-mapping.js call into amIAddonManagerStartup's `registerChrome` instead.

Notably, however, contentaccessible=yes directives for chrome URIs were being ignored
except for testing environments. That was originally added in bug 1451503 when the
reftest addon was ported to a WebExtension. I've gone ahead and removed the NS_ENSURE
check for being in the automation case.

Differential Revision: https://phabricator.services.mozilla.com/D240941

UltraBlame original commit: 707865adbdb7d308cecbd4c28498e388ec6824f3