Bug 1950424 part 4 - handle race-y condition with multiple requests and user action IDs r=dlp-reviewers,handyman
The new gtest added in part 3 was failing intermittently by hitting an
assert that indicated the userActionId and requestToken didn't match
what we had set up in MultipartRequestCallback.

The reason is that in DoAnalyzeRequest(), calling Send() on the client
will return the a response that does not necessary match the request!
Since the agent does not return a userActionId in its response (sigh),
we used to remember the passed-in userActionId and set that on the response,
but that is wrong.

Now we look up the userActionId from the mUserActionMap and set it on
ContentAnalysisRequest. This means that if the request is cancelled we
will not be able to find the real userActionId, so adjust
mUserActionIdToCanceledResponseMap to keep track of the canceled responses
by request token instead.

Differential Revision: https://phabricator.services.mozilla.com/D241092

UltraBlame original commit: 529ea124f966a7a382cf70e769c824aaa30b74de