Bug 1951190 - Put gradle 'build' folders in TOPOBJDIR for top-level builds. r=android-reviewers,adhingra
For consistency with the rest of the build system, have gradle build folders be
created in TOPOBJDR. Put the logic in shared-settings.gradle and apply it to all
projects. Existing gradle and taskcluster configuration uses the 'gradle/build'
prefix, so continue using that for now.

Only apply this when building top-level for now since a fair bit of automation
of fenix/focus/a-c projects expects build artifacts in the source directory.

Plugins are compiled as "composite builds" so they somewhat exist in their own
world. For simplicity, just manually specify the `buildDir` for each plugin.

Differential Revision: https://phabricator.services.mozilla.com/D240116

UltraBlame original commit: 2843981acabe6d7338008812153138c8ed7b86a1