Bug 1953984 - Fix servo version of KeyframesName::to_css (r=emilio)
The previous version required a "mut" annotation on the closure for
Servo no annotation for Gecko which is annoying to conditionally apply.

So it has been rewritten to use a plain function instead of a closure.

Differential Revision: https://phabricator.services.mozilla.com/D241574

UltraBlame original commit: d18bb7e8e6bdd8afb9f71b20936fc2a8513b9664