Bug 1953974 - Remove an invalid assertion. r=dholbert
In this case the center is positive, min and max are negative, so we
know the center is greater than both.

But we still can't compare the min and max because they're different
units, so it can't be simplified.

Differential Revision: https://phabricator.services.mozilla.com/D241722

UltraBlame original commit: 4907b0c9a735b131b41004b9d07d78b70506fa47