Bug 1955059 [wpt PR 51455] - Fix handling of infinity in late-resolved relative colors, a=testonly
Automatic update from web-platform-tests
Fix handling of infinity in late-resolved relative colors

When converting a CSSMathExpressionNode tree to a CalculationValue tree,
we check whether the expression can be evaluated to NaN or infinity, and
if so return an appropriate clamped value. That check was not accounting
for the potential for color channel keywords to appear in the expression
and instead implicitly treated any such keywords as having a value of 0.

The fix is not to check for NaN or infinity when the expression contains
color channel keywords. Instead, the evaluation happens once the channel
keyword values are known and can be substituted in.

Fixed: 404218611
Change-Id: Ieff3b60acf9989c1bee4d9139d5e56c761f94f7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6370426
Commit-Queue: Kevin Babbitt <kbabbittmicrosoft.com>
Reviewed-by: Rune Lillesveen <futharkchromium.org>
Cr-Commit-Position: refs/heads/main{#1434842}

--

wpt-commits: b619ac5b841c50080218fa3f5ab79ca62225281f
wpt-pr: 51455

UltraBlame original commit: 2a7401ff4d3883b0f980f8a4cd8caddaeb27d3fe