Bug 1931407 - wasm: Rework bytecode handling for compilation to avoid copies. r=yury
We have two problems in bytecode handling:
  1. When streaming compiling, we have dis-contiguous buffers for 'env' 'code' and 'tail'
     regions of bytecode that we parse. This is because while we're background compiling
     the code section we cannot resize that buffer, and so we have a separate 'tail'
     buffer for those bytes as they come off of the wire. However ModuleGenerator::
     finishModule requires the bytecode to be contiguous so we make a big copy of the
     whole module.
      - Then this copy is thrown away if we're not tiering.
      - For eager tiering, we actually create a new copy of just the code section and
        break the bytecode back apart to what we originally had.
  2. When we're doing a synchronous compile, we make a copy of the module bytecode from
     the original array buffer. However there's no need to do this and we could actually
     just decode directly from the array buffer memory.

This commit fixes (1) by switching all of our bytecode handling to work on the (env, code,
tail) version of bytecode and not assume it is contiguous. This is abstracted away in a
new BytecodeSource class which handles the details.

We can fix (2) by making BytecodeSource just have a weak reference to the bytecode, and
adding a BytecodeBuffer class which actually owns the buffers. When we sync compile, we
just use BytecodeSource and never construct a BytecodeBuffer.

Differential Revision: https://phabricator.services.mozilla.com/D235239

UltraBlame original commit: 68e60c65b23fb6b8c4daa0a42eb32f10b4910e86