Bug 1941595: Make nsMenuBarX::PaintAsync into a sync method for macOS versions macOS 14 and later. r=mac-reviewers,mstange
This is attempting to do an effective backout of Bug 1765391, which
added the PaintAsync method. That was the working part of that Bug, and
it was presumed to be the source of the grey/disabled appearance that
provoked *this* Bug.

Differential Revision: https://phabricator.services.mozilla.com/D241856

UltraBlame original commit: 913871d0f259d953c25aa9ecf42684f48feb18bc