Bug 1955697 - Simplify scrolled canvas background painting. r=mstange
Always paint the scrolled canvas background if it's CSS-specified.

Compute the canvas background upfront so we don't need to walk the
display list after the fact.

This removes the scrolled canvas background in a few cases where the old
code could prove that it was not needed due to blending, see the
comment:

    // To make layers work better, we want to avoid having a big non-scrolled
    // color background behind a scrolled transparent background. Instead, we'll
    // try to move the color background into the scrolled content by making
    // nsDisplayCanvasBackground paint it.

It's unclear if that matters for performance with webrender nowadays,
but removing it does cause significant fuzzy failures, so it will be
reintroduced in a more explicit way in a follow-up patch.

Differential Revision: https://phabricator.services.mozilla.com/D242017

UltraBlame original commit: 01cecf5bf7e5603d5d7ebe4c28d8fdc0a74e6b51