Bug 1899503 - P2 PRemoteWorkerDebuggerMangaer binding implementation. r=asuth
This patch implements the PRemoteWorkerDebuggerManager binding.

PRemoteWorkerDebuggerManager is the IPC that handles the registration/unregistration of remote debuggers.
Its life cycle is as long as the RemoteWorkerService such that it can also handle the remote debugger of child workers.
The detailed binding flow is

    When ContentParent calls StartRemoteWorkerService() on the parent process main thread, creating Endpoints of PRemoteWorkerDebuggerManager, then creating and binding RemoteWorkerDebuggerManagerParent to the corresponding parent Endpoint.
    Propagating the created child Endpoint through ContentParent::SendInitRemoteWorkerService().
    Propagating the created child Endpoint to RemoteWorkerService::InitializeOnMainThread(), then creating PRemoteWorkerDebuggerManagerChild and binding it to the child Endpoint. Then IPC is done.

RemoteWorkerdDebuggerManagerChild can also be created in the parent process for workers on the parent process.
This is unnecessary, but providing a general debugging mechanism for all types of Workers is good.

The details flow is similar to the normal flow, but creating Endpoints and RemoteWorkerDebuggerManagerParent when RemoteWorkerService::Observe() the "profile-after-change" signal. Then, RemoteWorkerService::InitailizeOnMainThread() will be called directly on the parent process main thread.

Depends on D230259

Differential Revision: https://phabricator.services.mozilla.com/D231682

UltraBlame original commit: 9bd2a82481f455019b488d0c7ce9cc2136281489