Bug 1928476 - Report validation failure telemetry earlier r=aminomancer,nimbus-reviewers,chumphreys
This is the first of a series of refactors to enable the removal the
"six hour enrollment gap" that occurs because we unenroll from unseen
experiments only after evaluating all experiments.

This refactors moves the reporting of validation failures out of the
`ExperimentManager` and into the `EnrollmentsContext`, which lets us
send much less data between the two objects when completing an update.

Co-authored-by: Shane Hughes <shughesmozilla.com>

Differential Revision: https://phabricator.services.mozilla.com/D241509

UltraBlame original commit: eda1942e17c37a04dd73817c0f4ae365dd29311c