Bug 1941595: Make nsMenuBarX paint into a sync call for macOS versions macOS 14 and later. r=mac-reviewers,mstange
This is attempting to do an effective backout of Bug 1765391, which
added the PaintAsync method. It defines a new method, PaintAsyncIfNeeded
which replaces the callsites formerly using PaintAsync. That was the
working part of that Bug, and it was presumed to be the source of the
grey/disabled appearance that provoked *this* Bug.

Differential Revision: https://phabricator.services.mozilla.com/D241856

UltraBlame original commit: 85b141632fb2d751980e0be51d0aecdcd4b4ca56