Bug 1829935 - Fix browser_103_cleanup.js r=sunil,necko-reviewers
The test assumes that the request will get cancelled before the actual
pageload, but that may not be a fair assumption.
Instead we should have the pageload not actually use the image in the
early hint. That should ensure that it always gets cancelled.

If there are any more failure, we should consider not using square.png
in early hints, as that could be used by other tests being run in
parallel.

Differential Revision: https://phabricator.services.mozilla.com/D240610

UltraBlame original commit: 0453f7947912402f380e105e1fcfefdfa4458bfc