Bug 1956188 - Fix updating chunk list after freeing more than one arena has been allocated r=sfink
This moves the count updates into the same place as the list updates so it's
not possible for them to use different information.

As suggested in code review ArenaChunk::unused is renamed to isEmpty. This fits
better with the chunk classification into one of empty/available/full.

Differential Revision: https://phabricator.services.mozilla.com/D242901

UltraBlame original commit: 3be1bacc0685b42978f684a19af4f2e243648495