Bug 1957016 - Extract the preferences and presets code into a separate files r=canaltinova,profiler-reviewers
background.sys.mjs is a big file, that's included only in Firefox desktop.
This patch extracts the preferences and presets code into a separate
file.

The preferences and presets code are very interleaved, that's why
they're kept in the same file instead of splitting them further.

This new file is also located in the "shared" directory of devtools, so
that it will be shipped to Android as well.

Differential Revision: https://phabricator.services.mozilla.com/D242690

UltraBlame original commit: 3df90e75d86091148b5af95b8f819104f6fd8135