Bug 1957166 - Support runtime GTK theme switching on modern gnome environments. r=stransky
On Wayland this now gets through via the settings portal. Just forward it to
the settings object which will do the right thing (trigger the settings-changed
signal and all our update machinery).

There's the question of whether we should try to read this from dbus as well
but given the cost and that it seems to not be needed, I'll skip.

Differential Revision: https://phabricator.services.mozilla.com/D243567

UltraBlame original commit: 34d3c99e207cf7a7bc343b63af0b9dd9f1675a9e