Bug 1955534 - Do not use AsyncShutdown.quitApplicationGranted in SessionStore. r=sessionstore-reviewers,dwalker
SessionStore wants to flush the window and tab status early enough to not be disturbed by things going down, that is in response to "quit-application-granted".
Previously we used the AsyncShutdown blocker quitApplicationGranted here, but that is deprecated as it's not a "real" shutdown phase and caused a loot of confusion elsewhere.
Given we already had our own timer to actually avoid the crash from AsyncShutdown, there is no point in using AsyncShutdown and we can just spin the event loop by ourselves here.

Differential Revision: https://phabricator.services.mozilla.com/D242658

UltraBlame original commit: 2e637aa1dfcd296a1f960797190f8bb7f454adf3