Bug 1957890 [wpt PR 51787] - [PEPC] Enter fallback mode when the type attribute is invalid, a=testonly
Automatic update from web-platform-tests
[PEPC] Enter fallback mode when the type attribute is invalid

To allow for backwards compatibility the permission element is no longer
a void element. This allows a site to add contents inside the permission
element which will only be visible if the browser does not support the
permission element, therefore providing the same overall html for both
browsers that support and don't support the permission element.

In addition, we want to mimic that behavior in the case that the type
attribute is not valid. This represents a browser that supports the
permission element but not the exact capability in the type attribute.

This CL adds a slot for a child nodes to populated. This slot is shown
if the permission element's type attribute is invalid. The permission
element stop behaving like a permission element and no longer adjusts
style, or trigger permission requests in response to clicks.

Fixed: 403209279
Change-Id: I5917296107c518bb5c12fb22ff931ff779aa1eaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6415799
Commit-Queue: Andy Paicu <andypaicuchromium.org>
Feels: Andy Paicu <andypaicuchromium.org>
Reviewed-by: Mason Freed <masonfchromium.org>
Cr-Commit-Position: refs/heads/main{#1441483}

--

wpt-commits: 9cc1a33a767edf765c40a79235d9db3bf5a8cc13
wpt-pr: 51787

UltraBlame original commit: 1ee9c5afe15302aa4645e32492bfe446aded1343