Bug 1956124 - Pass CompositorScrollUpdate over PCompositorController as a struct rather than loose fields. r=hiro,geckoview-reviewers,m_kato
This makes it easier to extend the struct in the future.

Also rename some related APIs to have CompositorScrollUpdate in their
name for consistency.

(It would be nice to rename GeckoSession.ScrollPositionUpdate to
CompositorScrollUpdate as well, but that's public API.)

Differential Revision: https://phabricator.services.mozilla.com/D242835

UltraBlame original commit: a763a1651248b0e24e9cad59f0608fd8903fdd67