Bug 1954597 - Part 7: Check for failed object loads in DocumentLoadListener, r=smaug,necko-reviewers,valentin
Without this change, if an object/embed element encounters a failed
load, such as a load of a 404 page, that error page's content is sent to
the content process, and then discarded by the object/embed element to
display the fallback content. With this change, the channel is instead
cancelled in the parent process, meaning that parent process logic can
more accurately verify that no content documents with bad principals are
loaded.

Differential Revision: https://phabricator.services.mozilla.com/D242424

UltraBlame original commit: 701582c6a844761ba0b9d6306e1feb907d58e632