Bug 1957367 - FS: Avoid calling ClientDirectoryLock::Id() off the owning thread in FileSystemDataManager; r=dom-storage-reviewers,jstutte
ClientDirectoryLock::Id() is currently safe to call off the owning thread
because it only accesses a const member. However, for consistency with other
quota clients and in preparation for an upcoming change to use
ClientDirectoryLockHandle (which enforces owning-thread access), this patch
avoids off-thread access to the directory lock directly.

Instead, we extract and cache the lock id on the owning thread immediately
after acquiring the lock, and store it for later use off the owning thread.

Differential Revision: https://phabricator.services.mozilla.com/D243655

UltraBlame original commit: 78ae23e3694acc3be4ecabf5228f6985693e32c7