Bug 1958070 - move telemetry handling out of BrowserGlue.sys.mjs, r=firefox-desktop-core-reviewers ,mossop
This moves all startup telemetry gathering into a separate file, to prepare
for some minor scheduling changes and avoid all the direct dependencies from
BrowserGlue.sys.mjs.

Note that the status quo in terms of scheduling (which this patch is not
changing, and the next patch in the stack only changes a tiny bit) suggests that
some of the scheduling of this startup telemetry is... somewhat haphazard.

We may wish to invest in better primitives to both avoid data collection taking
valuable resources/time on startup, but also offer better guarantees that the
eventual collection contains the data we need. Especially the "best effort"
scheduled items currently do not satisfy the second criteria. Either way this
should be easier with all the relevant code in one place and scheduled in the
same way.

Differential Revision: https://phabricator.services.mozilla.com/D244420

UltraBlame original commit: 466ded45dcd77c779952916beba9afc25a3274e6