Bug 1960779: Adjust viewBox-synthesized-in-img-001.tentative.html and its reference to omit width/height/viewBox attributes rather than setting empty values. r=firefox-svg-reviewers,longsonr
As discussed on the bug: empty attribute-values are mostly-harmless, but they
do trigger validation errors that get reported in the DevTools console and in
test logs. So it's better to omit the attribute entirely instead of setting it
to a bogus empty value, to avoid that console/log-spam.

(While we're at it, this patch also fixes up some over-indented lines.)

Differential Revision: https://phabricator.services.mozilla.com/D245815

UltraBlame original commit: 0df66c606392dcae364fc09834518f280b1110df