Bug 1953807 - [devtools] Display inherited rules from ::details-content rules. r=ochameau
This required a few changes both on the server and the client.
For such element-backed pseudo element rules, they need to be considered "closer"
to the selected element than the rules on the `<details>` element.
This means we also need to display those rules in their own dedicated "Inherited from"
section to make sure overridden declarations are properly marked as such.

A test is added to cover the different cases we have to handle.

Differential Revision: https://phabricator.services.mozilla.com/D245251

UltraBlame original commit: 2e20be07e39b9ce6811962b929e92d81cfc7ac51