Bug 1961115 - Pass new_frame_ready parameters via a struct. r=gfx-reviewers,gw
This patch is a preliminary refactoring to make it simpler to pass additional information to the C++ side and should not affect behavior.
In a followup patch the widget glue will have special handling for frames that need to be rendered but not presented.

Differential Revision: https://phabricator.services.mozilla.com/D245564

UltraBlame original commit: 9099f91a9ce6f0c563dfbdcf82202bbb99c4ad63