Bug 1960811 - Part 13: Remove LargeBuffer::marked flag r=sfink
Now the LargeBuffer object is stored in a small buffer we can use the containing
buffer's mark state instead of storing this separately. This small buffer does
not escape and the buffer allcoator is the only code that causes it to be
marked.

Differential Revision: https://phabricator.services.mozilla.com/D245876

UltraBlame original commit: 38b72f17858d320e848452e5000d4760420ea41a