Bug 1958328 - Add performance test for history semantic search r=perftest-reviewers,mak,places-reviewers,tarek,sparky
This patch introduces a performance test targeting the history using semantic search functionality. The test is designed to benchmark the efficiency and responsiveness of the history semantic search operations under various conditions. Shorter vs Longer search queries and Concurrent vs Non-concurrent inference when vector DB load happens as deferred task in the background. This captures latency, of the ML engine, end to end inference and also the semantic search for shorter and longer queries.
In addition to that it captures how much time does it take for the vector DB update process to complete for the 10000 history entries.

Differential Revision: https://phabricator.services.mozilla.com/D245087

UltraBlame original commit: 063ea853231638bcbd3cac501daa2ea9504431dd