Bug 1961011 - Improve storage of Java exceptions in glean crash pings r=android-reviewers,nalexander
We could have instead changed `messages` to contain the output of
`Throwable.toString()` (which includes the type name), however it's less
error-prone to separate the exception type explicitly. This changes the
java_exception metric to a different form, but this is okay from a
compatibility perspective as the data is simply stored as JSON.
Downstream consumers will need to be able to handle all formats.

Differential Revision: https://phabricator.services.mozilla.com/D246495

UltraBlame original commit: b2485c517ae88b8cbbc4301828286ea3f46acfe1