Bug 1961318 - Make `PointerEventHandler::UpdateActivePointerState` register synthesized mouse info if there is not r=edgar,dom-core
Despite the name, `sActivePointersIds` is a list of all pointers which is either
active or inactive.

`Element::SetPointerCapture` checks whether the given `pointerId` is known one
or not.  However, we register mouse pointer when entering over the document
and unregister it when leaving from the document.

However, they are handled when `eMouseEnterIntoWidget` and
`eMouseExitFromWidget` are fired, however, they are not fired automatically
if the mouse events are synthesized on a document which does not contain the
system cursor position.

So, we need to make the method register the `PointerInfo` when
`EventStateManager` receives `eMouseOver` or `ePointerMove` which is caused only
by mouse.

For making any `PointerEventHandler::GetPointerInfo()` callers can distinguish
whether the pointer is registered or updated by a synthesized mouse events,
`PointerInfo` should have the information as a `bool` member.

Differential Revision: https://phabricator.services.mozilla.com/D246027

UltraBlame original commit: 574356d50559d498a44ce43b895317ee45f00fcc