Bug 1956486 - Annotate some mixed content tests as failing after navigation. r=tschuster
As per the comments in the bug, this is already broken for <link
rel=stylesheet>. The issue is that we don't re-trigger the image request
now, as we cache the whole stylesheet.

An alternative would be to use a CSS variable or so to bypass the style
resource caching, lmk if you want me to do that instead.

Differential Revision: https://phabricator.services.mozilla.com/D243771

UltraBlame original commit: 8042121f2dcefb7699d518ff2909c54682c88cb0