Bug 1950511 - part 5: use the arch parameter with rpm packaging r=gabriel
The tool rpmbuild come from with the --target parameter which allow to build a
package for a particular architecture (if the building tools are available).

Since the repackage-rpm command do not build anything, this argument can be
used to register the correct information in the generated package.

The --target argument will also add automaticaly the architecture related
entries in the specification file.

Differential Revision: https://phabricator.services.mozilla.com/D241231

UltraBlame original commit: 2d837ca5679616eb28096ab1c65653f0dc23a3e1