Bug 1655503 - Remove clear search button from a11y tree. r=accessibility-frontend-reviewers,desktop-theme-reviewers,emilio,smaug
The clear search button is not exposed to assistive technologies
and is not able to receive focus. This is to ensure a positive user
experience when navigating through the search input.

Move prefs to default section in test/forms/chrome.toml

Differential Revision: https://phabricator.services.mozilla.com/D240396

UltraBlame original commit: 1647cabe49d7d60d23c6dba1897372208442130a