Bug 1958299 - don't retry tasks on SIGKILL. r=taskgraph-reviewers,Eijebong
retry-exit-status 137 was added during the migration from aws to gcp as it was
generally a symptom of spot termination.  Since then, worker-runner has learned
to detect when that happens and terminate the task as `worker-shutdown`, so
this should no longer be necessary when using a recent worker image.  It
also interacts badly with how d2g deals with exceeding maxRunTime.

Differential Revision: https://phabricator.services.mozilla.com/D246600

UltraBlame original commit: 089f4eaad6e6caa28517306ba26baf8ab15b3228