Bug 1963464 - Initial implementation of PaintTimingMixin for the PaintTiming and LCP entries r=sefeng,webidl
This patch implements PaintTimingMixin with only the `paintTime`, without the
`presentationTime`. The presentationTime property is nullable in the spec and
the browsers might choose to not implement it. We would like to implement it
in the future, but we will do that as a follow-up.

Differential Revision: https://phabricator.services.mozilla.com/D247305

UltraBlame original commit: 2c129757796067eb55c967382782cfb0fb6c3937