Bug 1869966 part 1: Add tests for existing IAccessibleSelectionContainer behaviour. r=eeejay
We did have support for IA2 testing when this was added.
However, because this interface is a relatively recent addition to the IA2 spec, it isn't included in the COM proxy dll included in Windows.
This means we can't run these tests without building and installing our own proxy DLL.
This has to be integrated into the Windows CI images, so we can't run these tests on CI for now and thus they're disabled on CI.
Even so, it's better to have local-only tests for this than no tests at all.

Differential Revision: https://phabricator.services.mozilla.com/D247871

UltraBlame original commit: bcecb7e3d8567e9f0e92dcba71d7fa12669a6b05