Bug 1963967 - Use PresShell::ScrollFrameIntoView instead of PresShell::ScrollContentIntoView for zoom-to-focused-input. r=emilio,botond
A big difference between ScrollContentIntoView and ScrollFrameIntoView
is that the former flushes pending layout. In the case of
zoom-to-focused-input, it works for an already-focused element, thus
when the element got focused, the document should have been flushed
via a FlushPendingNotifications call in
nsFocusManager::FlushAndCheckIfFocusable [1].

With ScrollFrameIntoView we can set the offset from the target frame so
that we can specify the cursor position to fix bug 1927600.

Though this change doesn't involve any user observable behavior changes,
it includes a mochitest to make sure that zoom-to-focused-input works
properly for an input element inside `content-visibility: auto` subtree
just in case.

[1] https://searchfox.org/mozilla-central/rev/7d73613454bfe426fdceb635b33cd3061a69def4/dom/base/nsFocusManager.cpp#2180

Differential Revision: https://phabricator.services.mozilla.com/D247525

UltraBlame original commit: 520682d4fb5e0f25bd1854b2d1355ade23ebda7b