Bug 1964154 [wpt PR 52284] - layout: Simplify `PositioningContext` by having it hold a single `Vec`,
Automatic update from web-platform-tests
layout: Simplify `PositioningContext`

`PositioiningContext` held two vectors, one inside an `Optional`, to
differeniate between the version used for a containing block for all
descendants (including `position: absolute` and `position: fixed`) or
only for `position: absolute` descendants. This distinction was really
hard to reason about and required a lot of bookkeeping about what kind
of `PositioningContext` a layout box's parent expected. In addition, it
led to a lot of mistakes.

This change simplifies things so that `PositioningContext` only holds a
single vector. When it comes time lay out hoisted absolutely positioned
fragments, the code then either:
 - lays out all of them (in the case of a `PositioningContext` for all
   descendants), or
 - only lays out the `position: absolute` descendants and preserves the
   `position: fixed` descendants (in the case the `PositioningContext`
   is only for `position: absolute`.)

It's possible that this way of dealing with hoisted absolutes is a bit
less efficient, but in general, the number of these descendants is quite
small, so it should not be significant. In addition, this decreases the
size in memory of all `PositioningContexts` which are created in more
situations everyday.

Signed-off-by: Martin Robinson <mrobinsonigalia.com>

--

wpt-commits: 932ea82851d0591298c9f8c772d987f7bebe45ff
wpt-pr: 52284

Differential Revision: https://phabricator.services.mozilla.com/D248801

UltraBlame original commit: 86dd4168e34b668145aad5ffe87b799d55400370