Bug 1959074 - Use a less-obviously-quadratic-time innermost-loop finder for loop unrolling.  r=iain.
The MIR loop unroller in bug 1950284 uses a somewhat simplistic algorithm for
finding innermost loops (for unrolling).  This commit changes it to use the
scheme shown at the end of BacktrackingAllocator::init().

This both simplifies the code and avoids the need to compare every loop in a
function against every other loop in that function, in order to establish which
ones are innermost.  It thereby removes a potential quadratic cost.

Differential Revision: https://phabricator.services.mozilla.com/D248931

UltraBlame original commit: e866bceefb803078e9a6e20e28eabda486246785