Bug 1467846: Part 5: Refactor JS::Dispatchable consumers to use a UniquePtr; r=arai,smaug
This patch is a major refactoring that introduces a concept of ownership of JS::Dispatchables via
UniquePtrs. Embeddings must now use UniquePtrs to reference Dispatchables passed by the engine.

Two static methods are introduced to Dispatchable: Run and ReleaseFailedTask.

Run is used to release the UniquePtr prior to running. We do this because we still rely on the
js_delete(this) call inside of OffThreadPromiseTask::run in cases of web assembly, namely cases when
we never dispatch to the embedding and need to clean up
[1](https://searchfox.org/mozilla-central/rev/6cc05c6620814bf3eb05278ccf3b1f49690b7740/js/src/vm/HelperThreads.cpp#1542).
This will be refactored in a future step covered by Bug# 1959032.

ReleaseFailedTask similarily releases the task. This relies on the task having been registered with
the OffThreadPromiseRuntimeState class. As the only override of `run` is in OffThreadPromiseTask,
thus far there have been no guarantees of how this is done. This will be addressed in the next
patch.

Depends on D212878

Differential Revision: https://phabricator.services.mozilla.com/D244936

UltraBlame original commit: 3cd70f69a47b86e345766e1b0b79f4a2a114cb44