Bug 1827914 - Merge some Baseline compiler and interpreter functions r=iain
When the bytecode will be shared (for now, just in the self-hosted case), it is
the same as the BaselineInterpreter bytecode, so we reduce code duplication.

One consequence is that scriptInternal() becomes less useful. If I keep it, I
need to add an implementation for BaselineInterpreterHandler which would be
identical to maybeScript(). Having three potential methods to get the script is
confusing, and we already have a pattern of using maybeScript() when
implementing combined compiler and interpreter methods.

Differential Revision: https://phabricator.services.mozilla.com/D245768

UltraBlame original commit: bdc3d0f631eb5a0cc2cbd89d02399ca5ea580213