Bug 1965691 - Improve panel animations a bit on Linux and prefers-reduced-motion. r=stransky,desktop-theme-reviewers,dao
The opacity animation is not motion, so should be fine to do
unconditionally.

The issue with Linux seemed to be the transform animation, so only
suppress that, rather than the whole thing. I did a smoke test of the
relevant WMs and they look fine. In practice we're not changing behavior
in the compositors affected by bug 1385079 (because transparency isn't
available there).

On Wayland if I animate the position I can reproduce the flicker from
bug 1800368 comment 6, but not if I only animate opacity, so I think
we're good. I can't reproduce the issues from bug 1800442 anymore.

Differential Revision: https://phabricator.services.mozilla.com/D248710

UltraBlame original commit: 8cade2d76b6c848b26ab00ed9cbf0ed718fecccf