Bug 1844878 - Eliminate redundant addProperty calls for wrapper preservation r=smaug,jandem
This patch replaces the use of the addProperty hook to preserve DOM wrappers
in favor of calling the wrapper preservation directly based on a class flag.
This additionally allows us to add an object flag on the shape indicating
that objects with that shape have already had their wrappers preserved,
avoiding redundant calls and allowing us to remain in the JIT for objects
with already preserved wrappers.

Differential Revision: https://phabricator.services.mozilla.com/D249143

UltraBlame original commit: f6ba67f510be8a9c11718442e06228212a1afe78