Bug 1966890 - Expand light-dark() to support images. r=dshin
See https://github.com/w3c/csswg-drafts/issues/10577 for discussion on
this. I'm hoping this gets eventually standardized but if not we can
always undo this and use if() or what not once that's available.

For now add an internal WPT reftest.

Differential Revision: https://phabricator.services.mozilla.com/D249788

UltraBlame original commit: 4dad885276cad43bf0c41b4464701384cd1af957