Bug 1948978 - Make `PointerEventHandler::CheckPointerCaptureState` cancel the new pointer capture if `lostpointercapture` event listener removes the new override element from the DOM r=smaug
`lostpointercapture` event listeners can update the pending pointer capture
element.  Then, if the pending capture element is becoming the new override
element is removed from the DOM, it should not start capturing the pointer
anymore.  Therefore, after we dispatch a `lostpointercapture`, we need to check
whether the pending capture element which was stored before dispatching
`lostpointercapture` event is still connected. Then, if it's not connected
anymore, it should just stop dispatching `gotpointercapture` on the removed
target and clearing the new override element which is set before dispatching
the `lostpointercapture` event.

Note that even applying this patch, there are some edge case issues [1][2].  I
think when it becomes clearer what should we do in the edge cases, we should
redesign `PointerEventHandler` for conforming to the new behavior.  Until that,
I believe we don't need to care the edge cases because of non-realistic
scenarios.

However, only with the change, the test does not pass. The reason is, the test
synthesizes `pointerMove` twice [3], but expects 3 `pointermove` events [4].
The last one is wrong.  When the pointer capture element is changed, pointer
boundary events should be fired but `pointermove` shouldn't be fired because of
no user input [5]. This issue is filed as [6].

1. https://github.com/w3c/pointerevents/issues/550
2. https://github.com/w3c/pointerevents/issues/551
3. https://searchfox.org/mozilla-central/rev/4bd620c218a01568d541453ce3b215ebf55bca63/testing/web-platform/tests/pointerevents/pointerevent_lostpointercapture_remove_setcapture_node.html#76,78
4. https://searchfox.org/mozilla-central/rev/4bd620c218a01568d541453ce3b215ebf55bca63/testing/web-platform/tests/pointerevents/pointerevent_lostpointercapture_remove_setcapture_node.html#86,88,92
5. https://w3c.github.io/pointerevents/#the-pointermove-event
6. https://github.com/web-platform-tests/interop/issues/955

Differential Revision: https://phabricator.services.mozilla.com/D248690

UltraBlame original commit: 90bfce73efda1b8909f17b54cf1f74227812949a