Bug 1965987 - Clone CookieJarSettings when inheriting from the parent context. r=bvandersloot,cookie-reviewers,anti-tracking-reviewers,necko-reviewers,valentin
This patch changes the way of inheriting cookieJarSettings from a
parent context to clone. This prevents the partitionKey of a given
context being changed by other loading contexts.

Also, we enfore the cookieJarSettings to be merged  whenever there
is a change on the cookieJarSettings. So, the child process can get
the up-to-date cookieJarSettings.

Differential Revision: https://phabricator.services.mozilla.com/D249963

UltraBlame original commit: a0aad46ef2ffef3107b0ce90f5a425eecc758a68