Bug 1966658 - Fix the test browser_user_agent_input.js. r=nchevobbe,devtools-reviewers
This patch fixes the issue that changeUserAgentFromSelector() doesn't
properly wait for the DevTools to reload.

Also, selecting the UserAgent option from the selector won't always
trigger a reload if the userAgent isn't changed. So, we introduce an
option to the funciton changeUserAgentFromSelector() to skip wating for
reload.

Differential Revision: https://phabricator.services.mozilla.com/D251145

UltraBlame original commit: 44ddf63788654904d02f853cac5885638fd52514