Bug 1943467 - wasm: Move FP ABI handling from builtin thunk to wasm callers. r=yury
Builtin thunks currently translate the system ABI around FP values
to the wasm ABI for FP values.

For ARM32 with the 'softFP' flag, FP values should be passed and
returned through GPRs. For x86, FP values are returned via the
x87 FP stack.

This commit moves these ABI adaptations to happen in the wasm
caller side, instead of in the builtin thunk. This matches how
masm.callWithABI works in JS code. Baseline and Ion code generation
needed to be refactored to track which ABI we are working with and
to be sure to handle the correct FP ABI.

To do this, an ABIKind enum is added to track the different ABI's
for calls. WasmBuiltin is temporarily added to describe the ABI
that builtin thunks use. With this commit, wasm builtins basically
use the system ABI with the exception that they still use the stack
frame layout that has the extra instance slots. So it's halfway
between 'system' and 'wasm'. A future commit will use the 'system'
stack frame layout and let us get rid of ABIKind::WasmBuiltin.

GenerateBuiltinThunk now takes an ABIKind that it uses for generating
the thunk. This is because we use this stub also for 'typed natives'
which must be the wasm ABI because they are directly imported into
wasm in certain situations.

Differential Revision: https://phabricator.services.mozilla.com/D244433

UltraBlame original commit: 33b66fdf90797f37c88eb4cee11dc4dd4df32b60