Bug 1969000: Change WPT naturalWidth-naturalHeight-unavailable.tentative.html to append its img elements dynamically with JS. r=TYLin
This patch doesn't fundamentally change the behavior or expectations of this
test.

This patch simply moves all of the img elements to a <template> (which is not
rendered), and then appends a clone of those elements to a container within
the body; and then those cloned img elements are what get tested.

This sets the stage for later patches to append additional modified copies of
these img elements, with small modifications (e.g. using 'srcset' rather than
'src').

Differential Revision: https://phabricator.services.mozilla.com/D251659

UltraBlame original commit: f4c23966e101532c2022dd9d5079026112a20da4