Bug 1963067: correct asserts related to SAB deletion and unregistered timeout tasks; r=iain,arai
I've done more testing with SAB deletion, and it was indeed broken. We needed to flip the condition
in the moz_assert. In addition, because we do not track TimeoutTasks using OffThreadPromiseTask, we
never register it with the PromiseRuntimeState. The result is that the assert checking if the
internal dispatch queue is not empty, the number of registered tasks must be greater than 0, fails.
I've made a quick fix for this, which checks if the task is registered or not and increments a
numUnregistered counter. I plan to revisit this with a better design in 1959032.

It is pretty ugly though so if you want me to make this better now I can.

Differential Revision: https://phabricator.services.mozilla.com/D250827

UltraBlame original commit: bef134fe99c2ae6a0b14a5540b92af46e82b897c