Bug 1969825 - Flush the previous pending pointerdown if it doesn't have a known interactionId by the time of next pointerdown r=sefeng
This issue seems like it's very hard to reproduce, and only happens
after keeping Firefox running after a few days. I noticed that it mostly
happens in the parent process, where it stops dispatching event timing
entries because of an unknown interactionId of a pointerdown event.

This seems to happen when a pointerdown event is fired but no pointerup,
pointercancel or contextmenu events are fired after that that match it.
I don't exactly know in which cases it happens, but I saw it happen in
the wild. If we clear the flush the older pending entries before
overriding them in the pendingPointerdowns map, we can solve this issue.

This is not in the spec, but Chrome already has this check. I updated
the spec issue to mention this in the spec as well once we rework the
pendingPointerdowns logic.

Differential Revision: https://phabricator.services.mozilla.com/D252109

UltraBlame original commit: 49df57a0056cdf352b9f0c09474dd8213a7c298c