Bug 1968932: Disable browser_textSelectionContainer.js for CCov builds. r=morgan
This is really unfortunate: it'd be good to capture test coverage for our IAccessibleTextSelectionContainer code.
Unfortunately, building the IA2 COM proxy dll with code coverage flags makes it depend on mozglue.dll, which can't be loaded in the context of the Python runner.
There doesn't seem to be a way to disable this for a single dll in the build system.
So, we have no choice but to disable these tests for CCov (code coverage) builds.

Differential Revision: https://phabricator.services.mozilla.com/D251803

UltraBlame original commit: 878a18c09a666ddbbc00b4736d21e66a6570b627