Bug 1956646 - Parse and serialize <pt-name-and-class-selector> for named view transition pseudo-elements. r=view-transitions-reviewers,firefox-style-system-reviewers,emilio
Extend the syntax of `<pt-name-selector>` to `<pt-name-and-class-selector>`
for named view transition pseudo elements.

Also, add a new parameter in the parser to make the difference between
the pseudo-element string and the pseudo-element selector.

The matching part is not implemented yet and we will do that in Bug 1964949.

Differential Revision: https://phabricator.services.mozilla.com/D251967

UltraBlame original commit: f578ff9ec78659e1093ec18b3472f8eb0f14654d