Bug 1859474 - Don't release TCPSocket on socket thread. r=necko-reviewers,kershaw
Since TCPSocket might be released on the socket thread, it might cause a
crash when TCPSocket is destroyed on this thread.

Since we use mozilla::SyncRunnable::DispatchToThread, this is synchronous
call. So it is unnecessary to touch reference count on this thread.

Differential Revision: https://phabricator.services.mozilla.com/D251990

UltraBlame original commit: b95312128b1446bc7a2d4bad4eaf47ba491f394c