Bug 1970605 - Add JSON storage checksum function to determine rolling migration checkpoint r=dimi
As part of the AS Login Project, we will have parallel stores for a
while to ensure data integrity in the new store: the JSON store as the
primary store and the new Rust store as a write-only mirror.

To do this, we will implement a rolling (continuous) migration to mirror
the data. However, since a profile can also be opened from a different
Firefox (e.g., a release version if the first FX is a nightly build), it
is conceivable that logins could be changed with an FX that has not been
migrated. To detect this scenario, we will store a checksum of the login
JSON in the Rust store and compare it when the logins are first
accessed.

This patch adds the corresponding functionality to the
JSONFile API, based on IOUtils, which does this in a performant
file-based manner.

Differential Revision: https://phabricator.services.mozilla.com/D252644

UltraBlame original commit: 86d763c23af5334ea3f9bfbbf9bb78fe17c57658