Bug 1923846 - Allow WebRenderBridge to use an extra pipeline ID. r=gfx-reviewers,lsalzman
This is quite the unsavory hack. For view-transition, the WR layer manager and bridge needs to be able to create a temporary pipeline without altering the one that is associated with the bridge and part of the scene that is being displayed. To do that the layer manager needs a pipeline Id to attach this premporary pipeline to, but allocating pipeline IDs is currently done in the main thread of the parent process and we can't afford the round-trip.
To work around this limitation, the patch makes it so that the pipeline IDs are effectively allocated in pairs. The even number is used for main pipeline ids, and the odd one can be used for temporary pipelines without risking collitions.

Ideally we would only allocate namespace ids (the upper 32 bits of the pipeline id) on the parent process main thread so that webrender bridges have the possibility to manage a large number of pipeline ids.

Differential Revision: https://phabricator.services.mozilla.com/D252500

UltraBlame original commit: ff548f5b2ab2a4966ba17892a3489935400500e1