Bug 1970980 - Always prefer to search for a color font for Regional Indicator codepoints, unless text presentation is *explicitly* requested. r=layout-reviewers,emilio
This makes us prefer our bundled Twemoji Mozilla font on Windows, so that emoji flags
render as users expect, unless the page explicitly asks for font-variant-emoji:text.

The test is a Windows-only in-tree reftest rather than a more general WPT test because
it is specific to the Windows font configuration, including the fact that we bundle
an additional emoji font. Behavior on other platforms and in other browsers  would be
too dependent on the local configuration for it to be useful.

Differential Revision: https://phabricator.services.mozilla.com/D253079

UltraBlame original commit: b700737cfc2ffc7b0e6e61550338f98906d1abce