Bug 1971630 - Skip subresource OnStateChange notifications for BCWebProgress, r=smaug
The changes from the previous patch to re-throttle OnStatusChange and
OnProgressChange appear to have not been enough. It appears that we are
also seeing a substantial amount of IPC overhead due to sending
OnStateChange notifications to the parent process for subresources, such
as images.

This re-adds some filtering (though it is still more relaxed than the
filtering was before bug 1963189) to OnStateChange notifications, so
that only notifications relevant to primary document loads will be sent
over IPC.

This particular filtering is done in the OnStateChange listener, as
listeners in the parent process may care about
STATE_IS_REDIRECTED_DOCUMENT, which cannot be filtered for at
registration time.

Differential Revision: https://phabricator.services.mozilla.com/D253411

UltraBlame original commit: f68866f34c9db0d812f13a6bb6970872f540518f