Bug 1969508 - wasm: Allow builtin modules to use the memory of their importing instance. r=yury
Allow a builtin module B that is instantiated by a module A to use the memory of module
A. This allows builtin modules that use memory, like "wasm_gemm", to be implicitly
imported like how "wasm:js-string" works. This then lets us optimize calls to "wasm_gemm"
to use the 'knownFuncImport' mechanism so that we remove the overhead of calling an
unknown import.

This works by:
  1. Track the import index for memories, if any
  2. When compiling a builtin module that uses memory, take an optional import
     descriptor that can be used for setting the import name the builtin module
     will look for.
  3. When instantiating a module that uses a builtin module, look for the first
     imported memory and use that import descriptor when compiling the builtin
     module. Then pass the module's imports object to the builtin module when
     instantiating it. Because they have the same import for memory, they will
     end up with the same memory object.

Differential Revision: https://phabricator.services.mozilla.com/D248489

UltraBlame original commit: 30774a2ebcfd9db233f5cce3e0fa5d038161150f