Bug 1965504 part 4 - don't show the normal Content Analysis dialogs for downloads r=dlp-reviewers,handyman
One annoying part to this is that for "Save As" downloads on a PDF (for
example), we don't get a valid browsing context. For downloads, the only
thing we need a browsing context for us to get a window so we can show
the "DLP busy" notification, so just use the window that we were
initialized with. In practice I can't tell any difference.

Differential Revision: https://phabricator.services.mozilla.com/D251880

UltraBlame original commit: 71ee118d056b4e822a1a93f4a9080897af9be5fc