Bug 1972259 - Skip view transition on page swap. r=nical
This fixes the crash, but it causes the commented out assertion to fire,
probably because the delete op is queued already with the wrong
pipeline, which I think means the snapshots leak.

Bug 1972525 is fixed to re-enable that.

Differential Revision: https://phabricator.services.mozilla.com/D253797

UltraBlame original commit: d0a241d0c9b60dd81a2de262111995582864544c