Bug 1972636 - Pass subsuite to mochitest when running perf tests in CI. r=perftest-reviewers,afinder
This patch resolves an issue where a mochitest test that is a part of a subsuite would not be found in CI even though the test exists. The subsuite is now detected automatically by parsing the test list that is returned from the `TestManifest` parser.

Depends on D253657

Differential Revision: https://phabricator.services.mozilla.com/D254051

UltraBlame original commit: 21566bf1045e9add157a4f5ad6e668c3b485c69f