Bug 1927886: Hold stream ids in a separate mainthread-only member, instead of trying to track them in a combination of mChannels and mPending. r=ng
This decouples the lifetime of the stream ids from the lifetime of the
channels, so they can live long enough to be properly reset. This fixes a
problem with stream ids being reused too early, because of the async steps in
the previous patch. It also makes it easier to extend their lifetime further
if needed, which is something that we'll need to revisit.

Also add a test for RTCDataChannel.id assignment.

Differential Revision: https://phabricator.services.mozilla.com/D238862

UltraBlame original commit: 882a6ac05ef014c6051ae512cd6f190c6129840d