Bug 1966426 - Create basic moz-promo component. r=reusable-components-reviewers,desktop-theme-reviewers,hjones
These changes create a new moz-promo element based on MozLitElement.

Currently this element utilizes Fluent for the 'heading' and 'message'
for this component. We may want to change this to 'heading'/'label' or
'label'/'description' in the future.

Utilizes two types, "default" and "vibrant", to change the colors of the
moz-promo element. We will need updated colors for both dark mode and
light mode, see moz-promo.css for TODOs.

Add stories for the "default" and "vibrant" types of moz-promo.

Add basic tests for the moz-promo component.

Differential Revision: https://phabricator.services.mozilla.com/D254108

UltraBlame original commit: 1edca3edf8361f7e1410169685fd96bc39b21f1a