Bug 1973631 - Do not cancel proxied transactions on LNA r=necko-reviewers,kershaw
The mPeerAddr used to check if a connection is LNA points to the
server address when using a proxy, not to the real website's IP.
That means we can't fail LNA channels when a proxy is involved
because we can't accurately detect if this is an LNA channel or not.

This patch excludes transactions whose connection info indicates
are using a proxy from being checked for LNA accesses.

Differential Revision: https://phabricator.services.mozilla.com/D254800

UltraBlame original commit: 7a053530b24b6d811fbfff4be16b8bfb9ae96e23