Bug 1951857 - tabs in collapsed tab groups should not be in accessibility tree r=dao,desktop-theme-reviewers,tabbrowser-reviewers
Added a `visibility` transition on tabs to match the transition duration that we use for the expand/collapse animation of horizontal tab groups. Added `visibility: hidden` on tabs in collapsed tab groups so that they are removed from the accessibility tree at the end of the collapse animation.

Differential Revision: https://phabricator.services.mozilla.com/D254326

UltraBlame original commit: de49c0429b24b8408171ea3cbc032a9abe217ee8