Bug 1972228 - Add some test coverage to keep data collection permissions in sync with gecko. r=rpl,jonalmeida,nalexander,geckoview-reviewers,android-reviewers,geckoview-api-reviewers,android-addons-reviewers,owlish
This patch allows us to keep the data collection permissions in sync
between Gecko, GeckoView and Android Components.

First, A new test helper has been added to the GeckoView layer so that
we can retrieve the list of permission names defined in the (toolkit)
WebExtensions JSON schemas.

In GeckoView, a new test guarantees that the set of data collection
permissions known by GeckoView (via the `DATA_COLLECTION_PERMISSIONS`
property) is the same as the one defined at the toolkit level. Since
data collection permissions might change over time, this will ensure
that we keep parity between Gecko and GeckoView.

The main reason for keeping all these permissions in sync is to make
sure Firefox for Android will get proper support, including localized
strings, etc. That's why the second part of this patch exposes the
GeckoView `WebExtension.DATA_COLLECTION_PERMISSIONS` property to the
`GeckoWebExtension` class in Android-Components, and a new **test-only**
dependency has been added to the _feature-addons_ component, which is
where localized strings for data collection permissions are defined.

Now I am not entirely sure about this, but I cannot think of a different
way to cover the list of localized strings without duplicating the list
of permission names like it was done before this patch.

I opted to add a new property to the _browser-engine-gecko_ so that we
don't need to deal with a GeckoView dependency in the _feature-addons_
(mainly because I doubt it's a trivial dependency declaration compared
to other libraries/components).

Differential Revision: https://phabricator.services.mozilla.com/D253784

UltraBlame original commit: 08f289ae511d8c1e81f7046458029745392904e1