Bug 1924030 - Fix verifyUBlockWorksInNormalModeTest UI test r=ohorvath
The UI test was flaky because it couldn't verify the number trackers blocked by uBlock.
This was happening because of the repeated calls to google.com (unusual trafic detected)

To fix this, I've changed the website that we're opening.

The UI test successfully passed 25x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D254947

UltraBlame original commit: f17da4b396034fad2a05de92946649106c98b18e