Bug 1953740 - part 2 - Extract functionality of configuring search suggestions providers r=android-reviewers,pollymce
Since AwesomeBarView contained all functionality for configuring
search suggestions providers and this logic would be needed in a
composable also the AwesomeBarView was renamed and cleaned to only
contain this logic.
A new class AwesomeBarViewSuggestions was introduced with the rest
of the configuration used by SearchDialogFragment for configuring
how to show search suggestions in a View.

Differential Revision: https://phabricator.services.mozilla.com/D250799

UltraBlame original commit: f362f362e19d8fb1e228d4137413d03500ecde89