Bug 1953740 - part 7 - Basic search middleware r=android-reviewers,pollymce
Extract the needed functionalities from `SearchDialogFragment` / `Controller
to a new middleware for `SearchDialogFragmentStore` that can drive a
composable awesomebar without the need of a separate dialog fragment in which
to show search suggestions.
Also refactored SearchDialogStore to be a UIStore to help with easier testing
and better align with the new application architecture directions.

Differential Revision: https://phabricator.services.mozilla.com/D250802

UltraBlame original commit: c01979d4f7314266423710b14da5ebfc99abd936