Bug 1973529 - UniFFI interface class names, r=markh
This implements what I think is the minimum for the bug.  Trait
interfaces that can be implemented in JS get `Impl` appended to their
name.

This removes the footgun that joschmidt ran into.  When implementing a
trait it's natural to extend the trait name with something like
`class JsPrimaryPasswordAuthenticator extends
PrimaryPasswordAuthenticator`.  However, this means you're extending the
implementation class that expects to make FFI calls and it fails in all
kinsd of weird ways.

With this change that code would still fail, but in an obvious way.

Differential Revision: https://phabricator.services.mozilla.com/D255320

UltraBlame original commit: ac0ab8d7dcd95643ca366a2e7f0523906bf3a7be