Bug 1947413 - Grid items should join the correct baseline sharing group r=dholbert
Calculate the correct baseline sharing group for grid items that are
baseline aligned. The calculation uses the algorithm from the spec here:

https://drafts.csswg.org/css-align-3/#baseline-export

In addition to corrently determining the baseline sharing group, the
baseline calculation can also be simplified.

Once we have the new baseline, we can use it in the alignment function,
which now has to take into account whether the baseline alignment
affected the side on which the item is aligned from.

For the additional test failure, it happens in subgrid and we use to
fail 11 of the 12 tests, now failing all 12, so likely we passed it by
accident.

Differential Revision: https://phabricator.services.mozilla.com/D253202

UltraBlame original commit: 8651a114b5e0e0cc1b441dab973385acd308687b