Bug 1968716 - automatic session restore should un-save groups that it restored r=sessionstore-reviewers,sfoster
With session restore enabled ("Open previous tabs and windows" setting), Windows and Linux appeared to be going directly to `restoreWindow`, which was reopening tab groups from the last session but not correctly removing those restored tab groups from the session's saved groups. The result was one copy of the tab group being open in the tab strip and a second copy of the tab group being in session state as a saved tab group. These two copies are independent and can diverge. This is a bad scenario that's more than a UI bug.

We made a test for this scenario in bug 1968716 and the test passes. However, the test has been calling `self.marionette.quit()`, which is equivalent to Cmd+Q or Ctrl+Shift+Q. The problematic behavior manifests when the user closes the last window by hitting the X button of the operating system window. After adapting the test harness to mimic closing the last window, the bug appears.

I updated the tests to exercise both avenues for closing Firefox.

Differential Revision: https://phabricator.services.mozilla.com/D254903

UltraBlame original commit: a3b75860ea756b0a1266388838f1e98d2614bcda