Bug 1973043 - if download was blocked on exit, remove it from the downloads list r=Gijs
If the DLP agent responds with WARN for a file, and the user quits Firefox
(after a confirmation dialog), we treat that as a block. Previously we
were just reporting that to the agent and deleting the file on disk. But
for safebrowsing, downloads that were blocked don't persist after a browser
restart, so we should do that too for DLP ones.

So we just delete the download, which includes a call to respond to the
DLP agent and finalize the download, but also removes it from the relevant
lists.

Differential Revision: https://phabricator.services.mozilla.com/D255212

UltraBlame original commit: 06ebd91587763dbe6125d7ea6b1e79aeb33ca5fa