Bug 1974911: Broaden some recently-introduced expected-timeout annotations to allow test-failure as well, for WPTs that are normally expected-failure. r=jmaher
All of these mac-specific annotations were added recently as automated updates
in bug 1973503.  They didn't account for the possibility that we sometimes
might not timeout; we need to allow for that or else we'll get intermittent
failures when mac is doing better than it was in the test-run that experienced
these timeouts.

Differential Revision: https://phabricator.services.mozilla.com/D255705

UltraBlame original commit: 884c12939b8470165aa0f958a3014603260ffe9e