Bug 1970339, part 2 - Text Fragments: Make timeout for creating text fragments global for all ranges. r=vhilla,mccr8
Before this patch, the 5 second timeout was applied
to each range individually. In some edge cases,
the number of ranges could be large. And even if
creating a text directive for a single range was
very fast, the accumulated time for all ranges
could exceed the 5 second limit.

This patch changes the watchdog to run
globally instead of per range, therefore
cutting off the whole process of creating
a text directive after 5 seconds.

It is not feasible to create a test case for
the timeout watchdog, hence there are
no changed tests.

Differential Revision: https://phabricator.services.mozilla.com/D252946

UltraBlame original commit: 6b507dc6e99b8b5ca1ed8d267134cd1fb7b48322