Bug 1974838 - Add the new option to dump retained display list from layout debugger. r=TYLin,fluent-reviewers,flod
The general and preferable way is to enable the preference,
`layout.display-list.dump`. However, it may dump too much information
because it dumps the log before and after painting. If we just want to
check the display list after painting (in a stable status), I think
dumping retained display list should work well.

Note that it may be empty if we call it before paint, or retained
display list is disabled. In general cases, it should be the display list
after optimization (because we paint the content already when you
click the menu item).

Differential Revision: https://phabricator.services.mozilla.com/D255566

UltraBlame original commit: 64deceec72a5ba70ccbc6c64f33832aa041e2ee9