Bug 1975409. Cleanup img element wpt test annotations. r=layout-reviewers,emilio
I was trying to get an idea of where we were with web platform tests with regards to img elements but I found too many annotations of android+fission and linux+x86 that made it hard to find the actual failures.

This is green on try, but we don't run any android+fission or linux+x86 tests in the complete normal set of tests.

Differential Revision: https://phabricator.services.mozilla.com/D255939

UltraBlame original commit: cfe2a63df8a1694ae8d04e43c1a1e0c67a1ef548