Bug 1966192 - Sort the frames with z-index order when capturing old state and new state. r=view-transitions-reviewers,emilio
An ideal way is to iterate the display list because they are sorted.
However, we add special view transition display items after capturing
the old state, so the display list is not suitable for capturing new
state. Therefore, we still have to sort the frame list in z-index order.

In order to minimize the list size for sorting, we only put the frame
which may create a new stacking context. this shouldn't have much impact on
the performance in the general cases.

Differential Revision: https://phabricator.services.mozilla.com/D255394

UltraBlame original commit: f0a32c9d6ab5464a678d7cec88eeddc2760465a7