Bug 1924026 - Fix verifyUBlockWorksInPrivateModeTest UI test r=aaronmt
The UI test was flaky because it couldn't verify the number trackers blocked by uBlock.
This was happening because of the repeated calls to google.com (unusual trafic detected)

To fix this, I've changed the website that we're opening.

The UI test successfully passed 50x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D255943

UltraBlame original commit: 5731cacd23d21414c8582182a2e007a6f4b7c1c0