Bug 1964600 - Introduce an IPC channel between child processes and the crash helper r=afranchuk
This channel is currently only used on Linux to rendez-vous with the
crash helper, obtain its PID and use it to enable the process to be
dumped when the Yama LSM is enabled. It will be used to actually request
a dump in the future, when the breakpad exception handler is removed.

Differential Revision: https://phabricator.services.mozilla.com/D254047

UltraBlame original commit: c9486da0aa0b91e203bd0ffc33d65f799f61cbee