Bug 1968423 - wasm: Simplify stack result handling in Ion. r=jseward
Ion stores stack results that are refs using MWasmStoreRef, and this
is only safe because we manually disable the barriers for the
instruction. Add a dedicated node that is safe for ref stack results
and just does the plain store we need.

Differential Revision: https://phabricator.services.mozilla.com/D255036

UltraBlame original commit: 86694580ac1d673b1124fff47060f52d6605dd5b