Bug 1966586 - Reuse other browser windows when opening _blank links in Taskbar Tabs windows. r=nrishel
This doesn't affect other tab additions, nor does it stop the tab bar
from appearing altogether. The idea is that _if_ another tab is somehow
made, the user should see it; but we should not create new tabs if we
can avoid it.

This also adds tests for opening URIs in popups and taskbar tabs to make
it less likely that this breaks in future.

Differential Revision: https://phabricator.services.mozilla.com/D253726

UltraBlame original commit: 9ccc6a2267cbf69c621fec973bd28573c2a45a1f