Bug 1973941 - Make NonMemMovable* checker aware of std::unique_ptr r=glandium
std::unique_ptr are always memmove()able if the default_deleter is used,
otherwise it depends on memmove()ability of the deleter.

std::unique_ptr is often used with incomplete types, which is fine but
prevent the analysis of those type. Be conservative in that case.

Differential Revision: https://phabricator.services.mozilla.com/D255030

UltraBlame original commit: e787cc53005fb466fc6c8b842360d64ba4793106