Bug 1940385: Skip 256180-2.html on win32. r=layout-reviewers,emilio
This test was annotated as skipped on win32 when it was added.  That skip
annotation was speculatively removed in bug 1895637, but as evidenced by the
intermittent failures, we do actually need it, so let's add it back.

Conveniently, there's already a code-comment explaining why we're skipping this
test for win32.  (That comment was left in by accident when the annotation was
removed, so it didn't map to anything for a little while, but now it usefully
explains the newly-restored annotation.)

Differential Revision: https://phabricator.services.mozilla.com/D257071

UltraBlame original commit: 52ffb145865fea9ae8c9f3ed7a0e121426956c7f