Bug 1950748 part 1: HyperTextAccessibleBase::ToTextLeafPoint: Descend to the end of an embedded object when given its end offset. r=morgan
Normally, an embedded object only spans 1 HyperText offset.
However, if the embedded object is the last child of its container, GetChildAtOffset(containerCharacterLength) will also return the last embedded object.
There are some rare cases where we end up with a HyperText offset that refers to the end of a container; e.g. the caret offset when clicking in the empty area at the bottom of a container after an input.
Previously, ToTextLeafPoint would descend to the start of the container in this case, which doesn't make any sense.
Instead, we now descend to the end.

In a subsequent patch for this bug, we will eliminate the case that triggered this behaviour.
However, I think this is still a correctness fix worth landing, as there may be other cases I'm not aware of yet.

Differential Revision: https://phabricator.services.mozilla.com/D256397

UltraBlame original commit: 9ef5afc383e75c435d7d2a3b4df5d45b675ce714