Bug 1973972 - [devtools] Wait for scriptReady flag in browser_document_tracer.js r=devtools-reviewers,nchevobbe
The test fails intermittently because it seems to trigger a click before the onclick handler was registered. Adding a flag and waiting for it from the ContentTask to make sure the script was processed seems to fix the issue.

Differential Revision: https://phabricator.services.mozilla.com/D256582

UltraBlame original commit: 910aee26362f269f0e4ca770479cd1bacb2adb2d