Bug 1976137 part 4: Remove DocAccessibleChild overrides of SendCaretMoveEvent and SendFocusEvent. r=eeejay
Instead of overriding these with intermediate functions that get and pass the caret rect, just get and pass it directly.
This avoids indirection and thus makes the code easier to follow and reason about.
I think this indirection was originally added because the need for the caret rect was previously Windows specific, but it isn't any longer.

Differential Revision: https://phabricator.services.mozilla.com/D256566

UltraBlame original commit: 66b4c64987b952431622f35d208af1c63fd9cdf7