Bug 1976876 - [devtools] Refactor CompatibilityActor#getNodeCssIssues. r=devtools-reviewers,ochameau.
The function was building multiple arrays to return a list of issues, which could
be costly.
We're now only build a list of declarations that we pass only once to getCSSDeclarationBlockIssues.
Since we're only checking the compatibility based on the property name, we don't
add declarations if we already added one in the array previously (instead of filtering
the list afterward to remove duplicated issues, like we were doing before).
Finally we extract a StyleRuleActor#parseRuleDeclarations method from what was
previously done in the StyleRuleActor form, and call that directly from getNodeCssIssues,
so we don't pay the cost for other data being computed in the form method, which
we wouldn't need here.

Differential Revision: https://phabricator.services.mozilla.com/D256988

UltraBlame original commit: d7b09c631cab11978324da4d93fe351000be1595