Bug 1933685 Part 5 - Concurrent Delazification: Only allocate relative indexes when needed. r=arai
Relative Indexes are added as part of the InitialStencilAndDelazifications
class, but they are only useful when the stencils are used for compiling script
using stencils.

This change make sure that we are no longer allocating them all the time, but
only keep them alive as long as one DelazificationContext needs it to walk the
scope chains.

Differential Revision: https://phabricator.services.mozilla.com/D248652

UltraBlame original commit: 9166109e54501d66a70d83ba929d57f244a1aeb2