Bug 1976960 - refactor(webgpu): work with copy of `aDesc` in `CommandEncoder::BeginRenderPass` r=webgpu-reviewers,nical
In preparation for remainder of the patch stack, clone `aDesc` so we have a mutable copy. We'll actually mutate it once we start accepting `Texture`s and not just `TextureView`s, which we'll convert in-place to views.

Differential Revision: https://phabricator.services.mozilla.com/D257053

UltraBlame original commit: e98afaa9f97b8847044dd904b60cf637d5fda5e6