Bug 1850817 - Change mpu files-changed to all of 'testing/' and 'python/'. r=perftest-reviewers,aglavic DONTBUILD
This patch sets the files-changed setting for the mozperftest unit test to run on any commit that has changes in `testing/` or `python/`. This should capture most of the changes that have broken the mozperftest unit test in the past but were not caught until later.

Differential Revision: https://phabricator.services.mozilla.com/D257702

UltraBlame original commit: 805e790de312834c07e14323e639c452be253196