Bug 1977386 - Make `HTMLEditor` related code treat the cases that "\n" is not a preformatted line break but is not collapsible white-space r=m_kato
There is a case that "\n" is not a preformatted line break, but it's
rendered as a non-collapsible white-space.  Therefore, for checking
whether a "\n" is a collapsible white-space,
`EditorUtils::IsWhiteSpacePreformatted` should be used instead of
`EditorUtils::IsNewLinePreformatted`.

Differential Revision: https://phabricator.services.mozilla.com/D257458

UltraBlame original commit: 30d6b34a5c58d67f46536f712900dfd7f83a0b9a