Bug 1820594 - Part 16-1 : Use sync version of LoadRequestedModules for chrome pages. r=jonco
In mochitest-browser tests, it's possible for the about:neterror page to
load. This triggers aboutNetError.html, which in turn loads aboutNetError.mjs.

Occasionally, the load promise from
LoadRequestedModules(aboutNetError.mjs) is never resolved because the
PromiseJobRunnable is not executed if the global is in the process of
shutting down.

To avoid this issue, we now use the synchronous version of
LoadRequestedModules when loading chrome resources.

Differential Revision: https://phabricator.services.mozilla.com/D257676

UltraBlame original commit: cd1f242dade1e315305450ed117660360c91a6d4