Bug 1773313 - Extract ELF file identifier from memory during shared library iteration r=mstange,profiler-reviewers
This commit modifies the dl_iterate_callback to extract ELF file identifier
directly from PT_NOTE segments in memory instead of always reading from the
file system. This makes sure that if Firefox is updated in between and shared
libraries are changed, we still read the correct in-memory library.

If we fail to get the ELF file identifier from PT_NOTE, we fall back to reading
it from disk again.

I avoided using the `FileID` code as much as possible. But still used
`ElfClassBuildIDNoteIdentifier` inside because this was going to be nearly
identical.

Differential Revision: https://phabricator.services.mozilla.com/D257816

UltraBlame original commit: 543a577471306c01348e4ffaf5f823cc6cb2272d