Bug 1977367 part 5 - Change value of TraceKind::GetterSetter. r=jonco
This is needed for the next patch because the nursery kinds need to come first.

In `Value::traceKind` we now need an extra check for symbols, but because
`JSVAL_TYPE_SYMBOL` and `JSVAL_TYPE_PRIVATE_GCTHING` are adjacent the compiler
will hopefully be able to optimize the common case with a single branch.

Differential Revision: https://phabricator.services.mozilla.com/D258441

UltraBlame original commit: 9eff9a1867fc960e210cc1b5d355581dc0241fd8