Bug 1978598 - Make 'prefix-env' feature more generic, r=jcristau,taskgraph-reviewers
Currently 'prefix-env' is meant to append something to an existing env
variable, e.g appending to $PATH. It does this by inserting the env at
the beginning of the command.

However, $PATH and other variables that we want to "append" to, aren't
the only things that we might want to interpolate as part of the
command. This renames the feature to "prepend-env" and makes it a bit
more generic such that instead of:

    prefix-env:
        PATH: "foo:"

You can do:

    prepend-env:
        PATH: "FOO:$PATH"
        BAR: "$SOME_OTHER_VAR"

Now both appending to an env and general env vars are supported.

Differential Revision: https://phabricator.services.mozilla.com/D258232

UltraBlame original commit: 8189eeca49b3dd43800c9bfdb22104d00339f751