Bug 1977641 - Add new arg to `./mach format` to explicitly allow skip checking if android formatters are valid in this context r=sheehan
With an upcoming change to `conditions.is_android` it will require `config.status` to exist. If it does not
exist `./mach configure` will be called. In the case where configure does not succeed, we explicitly fail.

Lando depends on `./mach format`, but Lando cannot guarantee that `./mach configure` will succeed (since it
depends on `./mach boostrap`, and we've had issues with bootstrap failing on the landing worker and preventing
all landings). We also can't do a try/catch, since it configure fails in the try/catch, it will always try again
on subsequent landings, wasting a bunch of time on every landing.

As such, it seems like the best option is to just not check for this condition in Lando, and to do that we'll
use an explcit arg. In Lando `conditions.is_android` always evaluated `False` anyway, so this isn't changing any
behavior, just making the upcoming patch compatible with Lando.

Differential Revision: https://phabricator.services.mozilla.com/D259154

UltraBlame original commit: 26f008e9be228b79b406e2bd86ddc3d754fb48cc