Bug 1979823 - Move PrivateBrowsingLockFeature constructor to where it happens r=twhite,android-reviewers
The PrivateBrowsingLockFeature lives longer that the activity it is
attached too, so lets keep the scope the same.

Added a `PrivateBrowsingLockUseCases` to follow our use case patterns
that replaces the `onSuccessfulAuthentication` which doesn't need to be
coupled together since it only dispatches to the `AppStore`.

Remove the components instance.

Differential Revision: https://phabricator.services.mozilla.com/D259019

UltraBlame original commit: 7f6b6ba39c8e04581cec7dae32399140b9e2ee4f