Bug 1977636 - [lint] Display source context in default formatter, r=linter-reviewers,sylvestre
Some linters include source code as context, often with little ascii
arrows that point at the problem. This can be invaluable for figuring
out where the problem is.

We have an attribute to store this information in the `Issue` class, but
unfortunately none of the mozlint formatters display it!

This gets the default formatter to attempt displaying the source, with
some basic colorization. The colorization is dependent on the format of
each linter, so it may need to be adjusted as we add different linters.

Differential Revision: https://phabricator.services.mozilla.com/D257531

UltraBlame original commit: 0afaa2157e02815477f8e5e157014beb212e8c6d