Bug 1966443: Skip NUL characters in text that is copied to clipboard or dragged r=gstoll,dom-core,edgar
Windows and GTK do not allow NUL characters in text data types that appear
on the clipboard or in DND payloads.  Mac and Android do not specify if this
is allowed but they use string types that permit it.  This patch normalizes
behavior across platforms by removing NUL ('\0') characters from text
flavors in those cases.

The test has exceptions for the behavior of different platforms when copying
a string that only contains NUL characters. Android and Wayland behave
slightly differently than others but it is not an important edge case.

Differential Revision: https://phabricator.services.mozilla.com/D258092

UltraBlame original commit: 1dc105b7f06555263d74b7704e161577f58c8551