Bug 1948756 - part 5 - Support voice and qr input in searches started in browser r=android-reviewers,harrisono
The new try-catch is a quick fix of an existing bug in InlineAutocompleteEditText
where immediately after gaining focus the autocompleted text is considered in
reporting the original text.
And this conflicts with the composable functionality which will try to focus when
the text is updated.

Differential Revision: https://phabricator.services.mozilla.com/D259421

UltraBlame original commit: f998ce28e9c2f51aca2b7254c23ff0c8d8bf2644