Bug 1970152 - Part 3. Implement support classes for MediaDrm provisioning. r=media-playback-reviewers,padenot
MediaDrm may require us to provision during EME playback, which consists
of sending a POST request to the given URL with a given sequence of
request bytes, granted by AMediaDrm_getProvisionRequest. When we receive
the response, we need to call AMediaDrm_provideProvisionResponse.

This patch provides the plumbing so that the decoding process is able to
request the content process to perform this transaction on its behalf.

Differential Revision: https://phabricator.services.mozilla.com/D259675

UltraBlame original commit: 7233f88826ade14472c92acf6ab69c318783f46b