Bug 1981133 part 5 - Support LoadValue in the transpiler and Ion ICs. r=iain
Bug 1977367 added nursery-values support so it should be possible to handle the
nursery case in the future if needed. What makes this a bit tricky is that we don't
know the `Value` type in that case in the transpiler so we'd either have to store
the type separately in the snapshot or in the `ValueOrNurseryValueIndex`, or load the
constant as an opaque boxed `Value`.

Differential Revision: https://phabricator.services.mozilla.com/D260029

UltraBlame original commit: db8bfed72bebe0e2893b11125a9f6a02dfaa376b