Bug 1981272 - Eliminate redundant addProperty calls for wrapper preservation r=jandem
This is nearly identical to the patch from bug 1844878 - it's just a backout of the
backout which made some minor adjustments due to new code that had come in. The real
fix is in the followup patch.

Differential Revision: https://phabricator.services.mozilla.com/D260062

UltraBlame original commit: a9d3718a4ee356342fd3ea14df06b639639fae5b