Bug 1959449 - Fix PerfStartup metrics not reliably getting fired by fixing how we determine app startup reason r=android-reviewers,kaya
Before now, in the ProcessLifecycleObserver of AppStartReasonProvider,
we relied on Handler().post() to be called only if an activity was being created.

In our investigations, this no longer holds to be true -- we noticed getting the reason being set to "NON_ACTIVITY" in the process lifecycle observer a lot of times, which means we log things like "unknown_main" in such situations.

To fix this, we modified the heuristics to use the importance to determine if it's a foreground process start. This is possibly not yet fool-proof, but it's our best bet so far, at determining if it was indeed launched for foreground interaction, or launched in the background (due to work manager maintenance operations or whatever).

Differential Revision: https://phabricator.services.mozilla.com/D260956

UltraBlame original commit: 54fc832b24b3e330b7a29e0d246f2430c6e9c597