Bug 1983605 - Part 6: Use a BitSet to store the waiting tasks r=sfink
This replaces a count and a linked list with a bitset of the task ids. We can
get the task pointers by indexing into the tasks array. And we can still check for
any waiting tasks with a relaxed atomic load operation.

Differential Revision: https://phabricator.services.mozilla.com/D261432

UltraBlame original commit: 081d7754bd54e43c83961075dcff4b9db0796c48