Bug 1983704 - Simplify AutoTArray guts. r=nika
While looking into the previous patch, I realized there's a scary
mismatch between what the is-auto bit means for rust's ThinVec and
Gecko's AutoTArray.

Rust's ThinVec assumes that the is-auto bit means that the buffer itself
is stack-allocated, but that's not correct: Gecko will set mIsAutoArray
for heap-allocated buffers that replace a non-heap allocated buffer.
This is used so that when shrinking the array we can go back to point to
the stack-allocated buffer. This is all a bit hacky as noted by the
various comments here.

The point is that, if we were to get here with one such buffer, we'd
leak: https://searchfox.org/mozilla-central/rev/1177f9aa639e8bbb6633441064ab9dbc846a1998/third_party/rust/thin-vec/src/lib.rs#1673

I think right now we can't even get there, except if somehow you
mem::swap or so a ThinVec, which is rather dangerous anyways as
explained in the Crate's documentation.

I'm not convinced that's worth preserving all that complexity, as most
AutoTArrays are temporary. Simplify how AutoTArray is dealt with, so
that the rust usage is sound.

I made AutoTArray::Clear() point back to the stack-allocated buffer
because it's an easy way of getting the old behavior back if some
consumer needs it... I'll of course run this through performance
tests...

Differential Revision: https://phabricator.services.mozilla.com/D257931

UltraBlame original commit: 9b0e0cbbbc4f0b9599c3722ec0a5db2ee003631c