Bug 1981861 - part 1 - New Window.setSystemBarsBackground to theme system bars even when edge-to-edge is enabled r=android-reviewers,mcarare
When edge-to-edge is enabled Android doesn't theme anymore the system bars
so we have to do it ourselves.
The new method allows configuring the window background for the status bar,
the navigation bar and any horizontal insets but it is state-less - calling
it again with a different colour will reset the backgrounds set before.

Differential Revision: https://phabricator.services.mozilla.com/D261095

UltraBlame original commit: 1feee6c62130119d9d0f207d67c1f3a00a3fa74b