Bug 1980094 - patch 1 - Create nsIFrame::RubyMetrics to return the em-normalized metrics of the default font. r=dshin
Our normal line-box metrics are based on font ascent and descent metrics,
but in some cases these may include a large amount of "internal leading"
(space above/below the main body of the glyphs, but within the declared
ascent and descent metrics of the font, for example to allow space for
stacked diacritic marks).

In such cases, positioning ruby annotations by "stacking" line boxes may
result in annotations that are a long way from the base text. A better
appearance (and more similar to Chrome's rendering) results if we base the
ruby positioning on the em-size, not on the (potentially much larger)
ascent and descent exposed by the font.

To support this, create a method nsIFrame::RubyMetrics that returns a pair
of metrics (ascent and descent) that have been "normalized" to sum to 1em,
effectively removing any internal leading built in to the font's metrics.

No behavior change, as the new method is not yet used anywhere.

Differential Revision: https://phabricator.services.mozilla.com/D259830

UltraBlame original commit: 037dad9a6a5533d4ffade2f89832ad6d1eacf8db