Bug 1980094 - patch 2 - Replace nsFontMetrics::EmAscent/Descent with TrimmedAscent/Descent, and trim internal leading equally from both sides rather than in proportion to the max metrics. r=layout-reviewers,dshin
Experimentation has shown that we get much better ruby-annotation rendering if we trim
the font's internal leading equally from the ascent and descent, so update the font
metrics methods that the nsIFrame::RubyMetrics method relies on.

There are currently no other clients of the nsFontMetrics::EmAscent() and EmDescent()
methods, so modifying them here does not risk affecting rendering more widely.

For better clarity, rename them to Trimmed{Ascent,Descent}, and update the implementation
to trim half of the internal leading from each of ascent and descent, instead of scaling
them in proportion to the metrics (as we do for the gfxFont::Metrics emAscent/emDescent
fields).

No behavior change, as nothing is yet relying on these metrics.

(We may want to reconsider how the gfxFont::Metrics fields are computed, too, but that
potentially affects other things as well so I don't want to touch it here.)

Differential Revision: https://phabricator.services.mozilla.com/D261380

UltraBlame original commit: aed72ea48d1b46f313ea5c996c676944534ce022