Bug 1971624 - Initial support of extension controlled config settings. r=reusable-components-reviewer-rotation,settings-reviewer-rotation,mstriemer
This patch adds support for extension controlled prefs using the new
config based settings. Note, when adding a new setting that is
controlled by a preference, you will need to use the store_id that the
ExtensionPreferencesManager is using to monitor the preference.

For example, if some extension requires Container Tabs and is controlling
that setting, the store_id needed would be "privacy.containers" as that
is the CONTAINERS_ENABLED_SETTING_NAME variable used by the
ExtensionPreferencesManager to monitor the setting.

This patch does not add the "disable extension" button that exists in
the current UI. This will be added in a follow-up patch.

Differential Revision: https://phabricator.services.mozilla.com/D260698

UltraBlame original commit: dd9c335571f31066a55612b89703fcda4e13ef47