Bug 1979968 - P2: Introduce TextLeafPoint::InsertionPointBounds. r=Jamie
As mentioned in the bug this method does not return bounds for special insertion points.

The bounds are probably correct 90% of the time, the exceptions are transition points in embedded bidi segments,
among other unforseen situations. These edge cases are things we can probably live with for the time being.

This addition also has a few changes:
1. HyperTextAccessible::TextBounds now returns bounds for identical start and end offsets.
2. TextLeafRange::Bounds now returns the insertion point for a collapsed range.
3. The browser caret rect test now uses the insertion point rect to compare against the caret rect.

Differential Revision: https://phabricator.services.mozilla.com/D259134

UltraBlame original commit: 70247197872294e47b0e430ffebba5f36908d216