Bug 198014 - Check for untransformed OP_SOURCE rectangles when drawing. r=aosmond
Sometimes we submit rectangles for drawing without a transform to bypass the transform.
However, for OP_SOURCE, we expected there to always be a transform to check when to clear
the background. This fixes the OP_SOURCE check to account for untransformed rects.

Differential Revision: https://phabricator.services.mozilla.com/D261805

UltraBlame original commit: 5c8b5f16988a83052e4671f399c66ad0a1833a78