Bug 1983168 - Avoid using gfxVars to determine whether Device.importExternalTexture() webidl is enabled. r=webgpu-reviewers,webidl,teoxoy,smaug
To determine whether this function is enabled we use the annotation
`[Func="mozilla::webgpu::Instance::ExternalTextureEnabled"]`. This
function currently checks the value of a gfxVar, which itself is
determined via a combination of a pref and blocklist status.

Unfortuntely using a gfxVar is problematic in certain tests which may
not have initialized gfxPlatform by the time we generate these bindings.

To avoid this, make this function only check the value of the pref
rather than the gfxVar. As a consequence, in cases where the pref is
enabled but the feature is blocklisted, the function will unfortunately
still be enabled. However, the C++ implementation does still check the
gfxVar value and will throw an error if the feature is blocklisted.

Differential Revision: https://phabricator.services.mozilla.com/D261704

UltraBlame original commit: 18bfdea19739423786b8b9d19ffa250cf2bb7e94