Bug 1965670 - Introduce a mechanism for the main thread to specify whether a smooth scroll should scroll the visual or layout viewport. r=hiro,layout-reviewers
For now, the main thread always requests smooth scrolling of the
visual viewport, to reflect that this what APZ actually does.

This is wrong for some call sites, and will be fixed in later patches
in this series.

Differential Revision: https://phabricator.services.mozilla.com/D261624

UltraBlame original commit: d06446442249ea376cdd9bf498debb9b2c7dc3fc