Bug 1833841 - make "close all windows" work more reliably r=win-reviewers,handyman
"Close all windows" sends WM_CLOSE messages serially to the open windows.
If we take too long to process the WM_CLOSE message, Windows will give
up and stop sending messages to subsequent windows.

In this case, getting the workspace ID was taking too long and triggering
this behavior. This change makes it so when a window is created, it queues
an update to get its workspace ID so mDesktopId won't be empty when we
try to close it, and so we won't do a synchronous update.

Differential Revision: https://phabricator.services.mozilla.com/D262153

UltraBlame original commit: 151998d5170f3edf015aa018d85d7fb10d7f1e6d