Bug 1981860 - Prevent dragged tab from moving beyond periphery r=tabbrowser-reviewers,sthompson,desktop-theme-reviewers
This appeared to bea regression with the bounds, but on closer inspection it was
related to the recalculation of the translate value to center the mouse over the
tab. This patch adds a further check to make sure the dragged tab is within the
bounds before recalculation of translate.

Side note, also updated endBound to only use periphery since it covers all cases.

Differential Revision: https://phabricator.services.mozilla.com/D260467

UltraBlame original commit: 778e336f80a6aba210f541ba3225e9bdbbc04a2c