Bug 1983894 - Don't include cookies in backups r=mconley,backup-reviewers
We have not tested the impact of including cookies in restored data.

There are several ways that a website's functionality could be broken by cookies being restored on a new machine. If we can test widely to rule these out as issues, then syncing cookies may be shown to be safe and desirable.

However, it is less risky for this feature, with its aggressive release timeline, to err on the side of caution.

Therefore, we will remove cookies from the set of synced data.

Differential Revision: https://phabricator.services.mozilla.com/D261646

UltraBlame original commit: 444a3d6d9382bc9928b595f827e3d450f89ed876