Bug 1984918 - Don't mark large nursery owned buffers twice r=sfink
This could happen with semispace nursery because a buffer can still be nursery
owned after it was marked if its owner is promoted but not tenured. This
resulted in us attempting to move an element out of a linekd list that didn't
contain it causing crashes.

The fix is to not trace the buffer containing the large buffer's metadata in
traceLargeAlloc, since this already done in both cases in
markLargeNurseryOwnedBuffer and markLargeTenuredBuffer. Also we need to check
we're not marking twice in markLargeNurseryOwnedBuffer which can be called on
other paths too.

Differential Revision: https://phabricator.services.mozilla.com/D262553

UltraBlame original commit: 2ef42e9968dbd65cf9e03c3cafdc757041a17910