Bug 1985382 - Add IONPERF=ir-graph support. r=iain
If IONPERF=ir-graph, IonPerfSpewer will now dump the iongraph.json
into a temporary file, and link to it from debug entries in JitDump.
Instead of mapping a PC to a line/column, we map it to LIR node ID's
so that the graph visualizer can use this to display sample counts.

IonPerfSpewer is moved from CodeGenerator to MIRGenerator and
lives next to JitSpewGraphSpewer. Methods are added to spew
passes to both of these. The JitSpew StartFunction/EndFunction
methods are also moved to MIRGenerator.

Differential Revision: https://phabricator.services.mozilla.com/D262676

UltraBlame original commit: dbb0c668af416ae5dfb4fbf153bce340ec7f3d13