Bug 1984004 - Ensure explicit popover invokers participate in focus fixups r=dom-core,smaug
nsFocusManager assumed that the invoker of a popover would always have
a `popovertarget` pointing to the popover, however this is not the case
as the invoker of a popover can be any element set via
`showPopover({source})`.

This change ensures that `showPopover({source})` sets a WeakPtr on
`source` to the popover - the `mAssociatedPopover` extended slot.

This also refactors nsFocusManager to use `GetAssociatedPopover()`
instead of looking up the element's popovertarget element.

Differential Revision: https://phabricator.services.mozilla.com/D263142

UltraBlame original commit: 8fced7a13c7b7f9ca6878ecafee7d5bcf74e3e3f