Bug 1983559 - Disable clang plugin inspection on instrumented builds r=glandium
Instrumented builds are just an intermediate step before actual build,
which are built from the same source.

If a bug where to be missed in an instrumented build as a result of the
absence of check, only the profiling code would run it, not an end-user.
And the code would be caught when compiling the code with the profiling
instrumentation.

In a non-cached setup, this reduces the instrumented build time by 17%
on linux workers.

Differential Revision: https://phabricator.services.mozilla.com/D261544

UltraBlame original commit: 3834c59af7a68ed2de253c4414a96b403890d36e