Bug 1986129 - Don't list addressbar-specific search modes in the unified search mode switcher for the new searchbar implementation. r=daleharvey
Ideally I think we would use the input's providers manager to get the active providers, but that's not how the search mode switcher is set up. It uses UrlbarPrefs to check which of the local search modes are available for the address bar, so for now just skip all that when the input isn't the address bar.

Differential Revision: https://phabricator.services.mozilla.com/D263681

UltraBlame original commit: 0e3714990a7b87260fc3e68c5360b5e9ed7b0dd4