Bug 1720570 - Add CreateCssModule; r=allstarschh
This adds a method to create a css module to the modules interface.
Unlike JavaScript or JSON modules, we don't have an ability to parse
CSS inside SpiderMonkey, so this method accepts an already parsed
CSSStyleSheet value from the caller, and wraps it in a module.

Differential Revision: https://phabricator.services.mozilla.com/D263410

UltraBlame original commit: 686f5d460d1ae4fd7aa5910f95dc92988fb07e4f