Bug 1982984 - Pre-expand file descriptor table at process startup on Linux. r=glandium
See the comment in the patch for details.  Chromium's version of this is
applied in the zygote after forking, and there's an argument for doing
this in the fork server now that it's enabled by default, but I think it
makes more sense in `main` after returning from the fork server hook.

I've tested this manually by using bpftrace to observe calls to
`expand_files`.

Differential Revision: https://phabricator.services.mozilla.com/D261842

UltraBlame original commit: 10452e48a71ab5f23370642a6822471f8e7c1ccf