Bug 1988118 - Move DoHController init and clearing the doh-rollout.* prefs to GeckoView:InitialForeground r=nalexander,geckoview-reviewers
[org::mozilla::fenix::components::Core::<init>](https://searchfox.org/firefox-main/rev/3de2d8eb788da636e08c85dbb7b453857d181e87/mobile/android/fenix/app/src/main/java/org/mozilla/fenix/components/Core.kt#196) initializes the `network.android_doh.autoselect_enabled` pref when the Nimbus experiment is active in order to initialize the DoHController.
The DoHController sets up aditional prefs like the user's region, doh-rollout.mode and doh-rollout.uri that indicate which DoH server to use when the user didn't manually set it up (network.trr.mode == 0). However, when the experiment ended, these prefs should be cleared because DoHController is no longer being initialized.

The problem with checking this pref via GeckoViewUtils.addLazyPrefObserver was that when the default was false, the handler would not get called.
Adding a `default: null` parameter made it so it was called when it was false too, however it was called twice during startup. First with a false value in response to GeckoView:ResetUserPrefs, then with a true value if the experiment was enabled. This meant that we'd clear the doh-rollout prefs at every startup which was inefficient.

This patch moves the DoHController initialization and clearing of the prefs to the GeckoView:InitialForeground handler. This makes it be called just once after the pref value stabilizes.

Differential Revision: https://phabricator.services.mozilla.com/D264553

UltraBlame original commit: 57f7f5793d9ba4ba9685b5c12698e48f8a62036a