Bug 1988833 [wpt PR 54830] - [animation-trigger] Switch IDL & CSS to action-behavior model, a=testonly
Automatic update from web-platform-tests
[animation-trigger] Switch IDL & CSS to action-behavior model

The action-behavior model replaces the behavior model (unfortunately,
"behavior doesn't mean the same thing in the 2 models).

In the action-behavior model, a trigger responds to the occurrence of
an action by exhibiting a behavior.

For event triggers, an "action" is essentially a UIEvent, e.g. click,
dblclick, etc. For TimelineTriggers, the actions are "enter" and "exit"
which, respectively, correspond to the entry of the specified entry
range and the exiting of the specified exit range.

A "behavior" describes what the trigger does to the animation in
response to an action. For example, "play-pause" means "if the
animation is paused play it, and if playing, pause it."

This patch implements a subset of behaviors proposed at [1].

[1] https://github.com/w3c/csswg-drafts/issues/12611#issue-3326243729

Bug: 441408561
Change-Id: If21e8377ca5c8d5ad7ce8f68edee4fc831737d48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6922990
Reviewed-by: Kevin Ellis <keverschromium.org>
Commit-Queue: David Awogbemila <awogbemilachromium.org>
Cr-Commit-Position: refs/heads/main{#1515094}

--

wpt-commits: 5aa642def1c645781d2a26dd9a61209dc258d140
wpt-pr: 54830

UltraBlame original commit: 80bde02e7c50251f19e2f5b0b25c9031bd022ed7