Bug 1987755 - Reduce the number of times of accessing the pref "browser.safebrowsing.provider.google5.enabled". r=dimi
Due to we now populate the list of the google5 provider, the
nsUrlClassifierUtils::ReadProvidersFromPrefs() will access the pref
multiple times and cause the test browser_preferences_usage.js to fail.

To Fix this issue, we reduce the number of access to the pref in this
patch.

Differential Revision: https://phabricator.services.mozilla.com/D265065

UltraBlame original commit: 4a73bf97414c0ffbbb07cfb41e0e2b7d6fe7fd0b