Bug 1979324 - Fix the insets of the Tab Manager and ensure the colors are correct for the status bar and NavBar r=android-reviewers,matt-tighe
Unfortunately, due to HomeActivity's default handling of Edge to Edge + insets, there was no way to have custom handling of the NavBar and status bar on only the Tab Manager. This required doing a workaround of reverting `TabManagementFragment` back to a `DialogFragment`.

This fixes the sizing issues we've been seeing in the Bottom App Bar. It wasn't really a "size" issue, rather the bottom nav bar was getting colored automatically to the same value as the bottom app bar and the FAB was very misaligned.

Differential Revision: https://phabricator.services.mozilla.com/D261486

UltraBlame original commit: 1705ed6b1d797a075e6d35cbe16bac312cd5bb4d