Bug 1989038: write `tests-by-manifest.json.gz` artifact in test loader r=taskgraph-reviewers,jmaher
This artifact is only needed by `web-platform-tests`, and only needs the contents related to those tests. The `resolver` that holds the data needed for this artifact is built up over the course of running the transforms for these tests, and is only available in whichever subprocess generates the tasks for the `test` kind.

Differential Revision: https://phabricator.services.mozilla.com/D265470

UltraBlame original commit: d74ef5b2c2010579b0bb6bcfca5691855ee030df