Bug 1988960 - Part 1: Simplify FinalizationRegistry by replacing registrations WeakMap with a HashMap with weakly held keys r=sfink
Currently we use a WeakMap for holding data on which targets were registered
with which unregister token.  Replacing this with a normal hash map and
sweeping the keys will allow us to see when unregister tokens are dying which
we need for the next patch.

This is also simpler and doesn't require creating an object to hold the record
vector, currently required because WeakMap values must be GC things.

Differential Revision: https://phabricator.services.mozilla.com/D265262

UltraBlame original commit: e08e867e0b6b3b9d3eb70206651ed89a0c9015c4