Bug 1988667 - part 1: Add WPTs to check the editing result around flex/grid items r=m_kato
I think that flex/grid items should be treated as blocks at least.  So,
the expectations look like the `<span>`s which are flex or grid items
are treated as the normal paragraphs.

Although they look like a table, so if we align the behavior to tables,
we should preserve the structure, but Chrome does not behave so and
the expectations are based on Chrome 142 except when the behavior is
completely odd or is not consistent between `flex` vs. `inline-flex` or
`grid` vs. `inline-grid`.

Differential Revision: https://phabricator.services.mozilla.com/D265210

UltraBlame original commit: 6717f4bd1cfcaa383f5a93fbbb931886d71c3ba8