Bug 1989652: move browsertime tests to their own kind r=taskgraph-reviewers,perftest-reviewers,jmaher,kshampur
This was a Claude associated change, but is virtually the same as what was done for mochitests in https://bugzilla.mozilla.org/show_bug.cgi?id=1982404. Desktop & Mobile would ideally be in their own kinds, but I chose not to, because it was not possible to do without changing `PREFIX_BY_KIND`.

Differential Revision: https://phabricator.services.mozilla.com/D265672

UltraBlame original commit: 7cb7b29bfad66c2c99d845d019ad7fb8ef3fa127