Bug 1990217 - Part 2: Move accounting for nursery owned buffers to the buffer allocator r=sfink
Previously we the accounting for nursery owned buffers happened in the nursery
while for tenured owned buffers this happened in the buffer allocator. This
moves the accouting for nursery owned buffers to the buffer allocator.

Accounting for malloced memory still happens in the nursery.

Differential Revision: https://phabricator.services.mozilla.com/D265796

UltraBlame original commit: 1fbafdc07d1495b00819b9201d07517f30da511b