Bug 1990217 - Part 3: Replace ownerWasTenured argument with nurseryOwned in the buffer allocator r=sfink
Currently we pass an argument 'ownerWasTenured' when marking nursery owned
buffers. This is inconsitent with the 'nurseryOwned' argument which is used
everywhere else in the buffer allocator but has the opposite sense.

Refactor to use 'nurseryOwned' everywhere to make this simpler.

Differential Revision: https://phabricator.services.mozilla.com/D265797

UltraBlame original commit: b236f6ed343029d28b722d88698532ff3508270b