Bug 1989107: Rewrite generateEnterJit for arm64 r=jandem
On arm64, the presence of the pseudo-stack pointer means that we would really prefer to allocate the stack frame first and fill it in, instead of doing a bunch of pushes. Instead of using the shared code, which does a ton of pushes, I rewrote an arm64-specific version of the equivalent code that doesn't have to wrestle with the PSP.

Note that in this case, we have enough registers available to keep the number of actual arguments live in a register between the underflow check and the point where we push the descriptor, so we don't have to read it out of the return value. I couldn't quite make that work in the shared code.

Differential Revision: https://phabricator.services.mozilla.com/D265900

UltraBlame original commit: f421a917f05f99c81da363f313faefa88b9e1e10