Bug 1988050 Part 2: Implement NativeLayerRemote request readback. r=ipc-reviewers,mstange,mccr8
This defines the actual message RequestReadback as a sync message. The
message must be sync because the callsites for snapshots use sync
semantics.

Once the message to take a snapshot has reached the parent process, this
patch asks the "real" native layer root to do the snaphot, using the
existing mechanism, then sends the pixels back in an Shmem. On the gpu
side, the Shmem is copied back into the buffer supplied by the original
snapshot request.

Differential Revision: https://phabricator.services.mozilla.com/D265635

UltraBlame original commit: e48311f9890c09766883e88f496333bdf974d6b2