Bug 1987945 - P2: Cache relations using multiple attributes for same relation type. r=Jamie
This allows us to calculate relations remotely when more than one
relation source is at play. For example, a details relationship will be
determined in order of priority: aria-details, commandfor, popover.

Knowing the relation source remotely allows further filtering. For
example, a popover relation should not be valid if the target is a
sibling of the button. We should determine that remotely instead of
recalculating and sending a new relation cache on each tree mutation
that might change the relation.

It also allows us to provide something like `details-from` in the
accessible attribute to let ATs know what the source of the relationship
is and for it to determine presentation.

The one sticking point in this change is that determining reverse
relations is less straightforward and requires us to interrogate each
target candidate to see if the relationship is mutual. This is because
relationship sets can be fully or partially "eclipsed" by a higher
priority source on the potential target.

For example:
<button id="toggle" commandfor="popover" command="toggle-popover" aria-details="details">toggle1</button>
<button id="show" commandfor="popover" command="show-popover">show</button>
<dialog id="popover" popover>popover</dialog>
<div id="details">details</div>

The cached stored reverse relationship for "commandfor" on "popover" is
both "toggle" and "show", but the direct details relationship of "toggle" is "details".
So the calculated DETAILS_FOR of "popover" should just be "show".

Differential Revision: https://phabricator.services.mozilla.com/D264636

UltraBlame original commit: b93b02b2427123e6680520cd184188d276b18153