Bug 1962089 - Ensure connection is removed from acrive conns, r=necko-reviewers,jesup
Instead of playing whack-a-mole, a better approach is to add a WeakPtr in connection that tracks its ConnectionEntry.
When the connection is added to ActiveConns, set the owner entry. This guarantees the connection is removed from ActiveConns in nsHttpConnectionMgr::OnMsgReclaimConnection.

Differential Revision: https://phabricator.services.mozilla.com/D266674

UltraBlame original commit: c7174d383c136aa85d3e5050224b11074d906043