Bug 1982170 - Unit test to verify threadsafe access via nsITimerManager::GetTimers r=xpcom-reviewers,emilio
nsITimerManager::GetTimers() calls nsTimerImpl::GetName(), which used to
not be threadsafe when the callback target is on a different thread.
This was fixed in patch 1 of bug 1979013, by eagerly fetching the name
at the initialization of the timer.

This patch provides the test coverage to serve as a regression test.

Differential Revision: https://phabricator.services.mozilla.com/D260612

UltraBlame original commit: 68f1a065f7a1fcbbd1a82a88a008608d00887c1c