Bug 1822531 - Fix verifyAddressAutofillTest UI test r=aaronmt
The UI test was flaky when trying to set the text to the "Addresses" fields.

This was happening because we were dismissing the keyboard to have all fields in view.
By doing this in some cases it actually triggered a "go back"  navigation existing the sub setting,
The keyboard dismissal isn't actually required, so I've removed it (most likely it was a leftover from when we ran the tests on the Pixel 2 which is a smaller device)

The UI test successfully passed 50x on Firebase ✅
All UI tests that use the fillAndSaveAddress function successfully passed 3x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D267214

UltraBlame original commit: 0f79686f04d90b456ad28a8720ad59b6f56baf39