Bug 1986711 - Closed menupopups shouldn't have an intrinsic size. r=layout-reviewers,dshin
We already avoid most of the reflow work during Reflow(), but this
code-path can still be expensive. We get it to try to compute the
hypothetical position of a popup, which arguably we should never need to
begin with, but this seems trivial to do here.

Differential Revision: https://phabricator.services.mozilla.com/D267288

UltraBlame original commit: 5af6809b76e2af42424e4a077af57cd6a52fc740