Bug 1988445 - Re-enable test_ext_runtime_getContext.html on android builds. r=robwu
This patch is re-enabling the test_ext_runtime_getContext.html mochitest
on android builds, on top of the changes applied from Bug 1876317 patch
(linked to this revision as its parent).

The perma-failure hit by test_ext_runtime_getContext.html seems to be due
to a pre-existing issue hit by TestRunnerActivity, which has been made more
immediately visible by the changes recently applied by bug 1987325.

Investigating the failure more closely confirmed that it was related to the
crash mentioned in Bug 1988445 comment 1 as initially suspected.

Unlike other test failures linked to Bug 1876317 in this case the reason for
the test failure was not due to the fact that refreshExtensionList was clearing
all the ExtensionWrapper instances on the TestRunnerActivity side, but a race
between the time needed for refreshExtensionList to get the update list of
extensions metadata (which was async on both GeckoView and Gecko sides) and
the call to the onBrowserAction method from the TestActivityRunner
ActionDelegate (originated from the browserAction onManifestEntry
lifecycle method), which was hitting a crash due to onBrowserAction
method trying to access the browserActions property on a non existing
entry of the mExtensions HashMap (as it was not yet created by the
refreshExtensionList async logic).

Depends on D265430

Differential Revision: https://phabricator.services.mozilla.com/D265431

UltraBlame original commit: 4b993f17cb6bbafd3d0b05f071ab9a1303e1d140