Bug 1991502 - Ensure we retry launching the GPU process after early launch errors. r=gfx-reviewers,bradwerth
This patch makes us restart the GPU process after it fails to
launch during the initial setup. This is prior to any IPDL actors
are setup, so there is not much interesting done yet from a
graphics perspective. As such we treat this as an unstable process
and follow the same fallback algorithm.

Differential Revision: https://phabricator.services.mozilla.com/D266701

UltraBlame original commit: 9ce97b4425e61ecd96eef71f7f7f8b4316100a32