Bug 1955947 - Part 1: Implement Document::HasBeenScrolledSince, replacing Document::HasBeenScrolled r=farre,layout-reviewers,emilio,hiro
This change replaces the `Document::HasBeenScrolled` method, introducing
`Document::HasBeenScrolledSince(lastGeneration)`. This method delegates to
PresContext, which keeps a record of the last scroll "generation" - an
integer counter incremented on each scroll. This is sufficient to determine
whether we should adjust the scroll during a navigation intercept: if the root
scroll generation has been incremented between the time the navigate event
fired and the time the intercept settles, it's likely that was intentional and
we shouldn't try to perform a scroll.

This change is required because Navigations that cause scrolling should
(such as prior .scroll() calls, or navigating to an anchor and
navigating back) should properly restore scroll position, but should not
do so when there is an explicit intent to scroll during intercept.

Differential Revision: https://phabricator.services.mozilla.com/D266430

UltraBlame original commit: 9d1929e60c1d733bffb2ab211e4a63bd98e7e912