Bug 1968890 - Part 5: Add ModuleLoadRequest::SetErroredLoadingImports(). r=jonco
To indicate the loading error for a top-level of the module from the dynamic import
is successfully fetched and parsed, but one of its dependencies fails to fetch
or compile.

//Example
import("a.mjs");

// a.mjs
import "./404.mjs";

This patch allows the ModuleLoadRequest::IsErrored() for "a.mjs" returns
true.

Differential Revision: https://phabricator.services.mozilla.com/D264807

UltraBlame original commit: 584beaae1d20fcf1f066f802e6412673028fa4e5