Bug 1947324 part 6 - avoid using Windows App SDK on opt-talos-xperf runs r=bobowen,perftest-reviewers,kshampur
On opt-talos-xperf, the Windows App SDK fails
calling DCompositionCreateDevice3() with ERROR_ACCESS_DENIED,
and the code assumes it is going to succeed so it proceeds to crash
deferencing null.

We haven't seen this crash anywhere else, and so I suspect this has
something to do with the way Firefox is being run there. So add a pref
to turn off using the Windows App SDK for just this test.

If this turns out to cause crashes in other situations, we should figure
out what the situation is and try to fix it.

Differential Revision: https://phabricator.services.mozilla.com/D267040

UltraBlame original commit: 53aa44de71d67d208a00cae329749c3738cda7ac