Bug 1992856 - Disable remote canvas as the final fallback step before disabling the GPU process. r=gfx-reviewers,lsalzman
Given that we would otherwise disable remote canvas when we disable the
GPU process, it is not meaningfully worse for users to just disable
remote canvas keeping the GPU process as the final step. It is still
preferable to losing the GPU process sandbox. This includes D2D canvas,
accelerated canvas and WebGPU, as well as WebGL on Android.

Differential Revision: https://phabricator.services.mozilla.com/D267667

UltraBlame original commit: a3c710bb3f6e0809e1244e7e88f1767af1845010