Bug 1990881 - basic intent ask vs search vs navigate heuristic r=Mardak,firefox-ai-ml-reviewers
building on bug 1989035 doing chat, sometimes we want it to search instead. we can do some very basic heuristic to set up an async api for it to be likely replaced by a engine/model

to start, this could reflect text of the submit button based on the typed text. eventually this might be reused for classifying quick prompts and suggestions

Differential Revision: https://phabricator.services.mozilla.com/D267500

UltraBlame original commit: 1597c90a9b629826def6a46e3997ca8810b10545