Bug 1937766 - Make sure pragma:no-cache is respected even when a cache-control header is present r=necko-reviewers,kershaw,emilio
Even though the pragma: no-cache header is deprecated according to
https://httpwg.org/specs/rfc9111.html#field.pragma this header is often
present on the web, sometimes conflicting with the cache-control header.

Normally cache-control would take precedence when there's a conflict,
but since other browsers treat pragma: no-cache as a definite signal to not
cache, a different behaviour risks webcompat issues.

We can probably make an exception for cache-control: immutable. Since this is
a signal that the resource is unlikely to change, it's probably safe to treat
resources as cacheable in this case.

Differential Revision: https://phabricator.services.mozilla.com/D266972

UltraBlame original commit: d6c67f10a8d58a19cce381b59652e07d3c593c11