Bug 1988716 - Part 2: Add ProcessGet() function to labeled metrics. r=chutten,toolkit-telemetry-reviewers
This patch allows labeled metrics to easily and in a standardized manner label themselves according to the process they're recorded in. It does this by adding a base a ProcessGet() function that automatically creates a label for the right process.

In the process I've also cleaned up the code to use a single template rather than a set of specializations to increase code re-use, and using compile time evaluations to ensure no additional code is generated.

Differential Revision: https://phabricator.services.mozilla.com/D264976

UltraBlame original commit: 3fb488e4c3d20015a6bacfb19f0ffd2b16f93d32