Bug 1992509 - feat(mach): add `--vcs-diff` to `mach vendor rust` r=supply-chain-reviewers,glandium
Add a `--vcs-diff` option to `mach vendor rust` that prints a diff of
"uncommitted" changes after the main vendoring operation. On success,
this is expected to be empty.

This will be consumed in a subsequent patch, where we upload this sort
of diff to Taskcluster jobs so that contributors can use it to fix their
patches and move on.

Differential Revision: https://phabricator.services.mozilla.com/D267478

UltraBlame original commit: f4cfc606d0cca5e384d139db80c70195d0187234