Bug 1990703: Don't call getAccessible in AccessibilityUtils.isKeyboardFocusableSpinbuttonSibling. r=ayeddi
Calling getAccessible forces a refresh tick in some cases.
Forcing a refresh tick allows accessibility tree updates to run.
This might cause problems if the Accessible we're checking gets recreated for whatever reason, since that will cause the Accessible to die while we're still querying it.
Ideally, we'd avoid recreating it in the first place, but this might be tricky to diagnose.
It's probably best to avoid extra refresh ticks if we can anyway.
Instead of querying the sibling Accessible, we just query the sibling DOM node directly.
This is sufficient for our purposes, albeit less generalised.

Differential Revision: https://phabricator.services.mozilla.com/D267904

UltraBlame original commit: 085575ff6ca5bc68468aa182a1fbf4c94a532183