Bug 1992811 - Enable important-dates suggestions for en locales in DE, FR, and IT. r=daisuke
This enables Suggest and important-dates suggestions for `en` locales in the EU
countries. For these region-locales, Suggest itself will now be enabled, but
every feature of it will be disabled except for important dates. So, no AMP, no
Wikipedia, no settings UI, etc. IOW it will only be an implementation detail
that important dates are implemented using the Suggest platform.

I changed the default value of the `settingsUI` pref so that the settings UI
will be hidden by default. Right now, the default is to show the full UI. If I
didn't change the default, I would need to make more changes to how
`QuickSuggest` initializes default prefs (the `SUGGEST_PREFS` const) so that EU
countries would have branching logic for the `settingsUI` pref depending on the
locale, which is not currently possible.

Currently there isn't any data in remote settings for these region-locales, so
even with this patch, important dates won't be able to be triggered. I'll upload
some data later.

Depends on D268042

Differential Revision: https://phabricator.services.mozilla.com/D268044

UltraBlame original commit: 04f1335cc821d299065145b9340183a4ff82d68c