Bug 1993371 - Update region- and locale-handling in Suggest tests. r=daisuke
This rewrites `QuickSuggestTestUtils.withLocales` with the following goals in
mind:

* Wait on `SharedRemoteSettingsService` to update itself due to the
  region/locale change (bug 1966166 D268036).

* Re-initialize Suggest (by default) so that the Suggest prefs are properly set
  depending on the new region/locale.

* Force sync so that the Suggest store is populated with RS data for the new
  region/locale, so that callers don't need to do that.

I also factored out and reworked some code so that hopefully it's easier to
understand overall.

I split `withLocales` into `withRegionAndLocale` and `setRegionAndLocale`.
Callers can use `setRegionAndLocale` when they don't need to reset the original
region/locale. The motivation for this is the important-dates test: In
bug 1992811 I'm going to enable important dates for en locales in EU countries,
and I want to test a variety of regions/locales. By default, the test assumes
the US region and the en-US locale, and I use `setRegionAndLocale` in the setup
task for that.

Depends on D268036

Differential Revision: https://phabricator.services.mozilla.com/D268042

UltraBlame original commit: b3e4d8a43bdd0725bc21a4c4319c46ef98c87e05