Bug 1933181 - Stop using nsView for popups. r=layout-reviewers,tnikkel
Make the popup frame listener to the widget events itself, rather than
via the view tree.

This is a first step towards getting rid of views as a whole.

This creates the (annoying, but temporary) subtlety of the view tree
jumping across popups, so nsIFrame::GetNearestWidget and co need to
become a bit careful and avoid doing that.

Similarly a couple places relied on view visibility, which now need to
check nsMenuPopupFrame::IsOpen() as well.

Other than that this seems to just work.

Differential Revision: https://phabricator.services.mozilla.com/D230106

UltraBlame original commit: 832b10027e3b1c44f0a22891903f657bf2a97064