Bug 1994358 - Call SetTransparencyMode on popup creation. r=tnikkel,layout-reviewers
The regressing bug removed this call, effectively:

  https://searchfox.org/firefox-main/rev/644f0db17749554fe23a45b43e77e61f42acdfd9/layout/xul/nsMenuPopupFrame.cpp#340

When I removed, I did so because it should be redundant (the
transparency mode is passed in through InitData). However it is not,
because only GTK looks at InitData::mTransparencyMode.

For now, preserve behavior exactly. Long term we should rejigger stuff
so that BaseCreate() or so calls SetTransparencyMode(), but that
requires testing all platforms since that gets called in a
partially-initialized state.

Differential Revision: https://phabricator.services.mozilla.com/D268691

UltraBlame original commit: 7e1119aaa2351c73691c02d23794affc58310535