Bug 1967475 - part2 : check the first sync sample in MP4 if a key frame can't be found from the closest sync sample. r=media-playback-reviewers,jolin
Due to D268192, now only IDR frame will be marked as keyframe on MacOS.

That means other type of I frames will be ignored, in this situation,
seeking from the closet sync sample might not be able to find a key frame,
because the sync sample might be other types of I frames.

Therefore, we should try to search from the first sync sample in a GOP, which
is usually a IDR frame, in order to find a valid key frame.

Differential Revision: https://phabricator.services.mozilla.com/D268213

UltraBlame original commit: 9f2387fd7d4851be0158d220fe80068d47836020