Bug 1993198 [wpt PR 55287] - [gap-decorations] Flex fragmentation gap decorations, a=testonly
Automatic update from web-platform-tests
[gap-decorations] Flex fragmentation gap decorations

This CL implements gap decorations for flex in fragmentation scenarios.

Essentially what we do now, is we build a `GapGeometry` for every
fragment, making everything within that geometry fragment-relative.

In order to achieve this there was some minor refactoring to the
`GapAccumulator` logic so that the only thing we have to do is pass in
the fragment-relative offsets and booleans so that we build the Geometry
correctly. For example, the accumulator uses boolean states to build the
geometry, such as whether we are processing the first or the last line.

This CL does some refactoring so that that same logic applies but is
fragment relative (i.e. whether a line is "first" in the fragment).

Now we will create a GapGeometry in `GiveItemsFinalPositionAndSize` when
there's no fragmentation, and in
`GiveItemsFinalPositionAndSizeForFragmentation` when there is
fragmentation.

Tweaks for column flexboxes as well as their respective tests will be
added in a followup CL.

Bug: 357648037
Change-Id: Ieb5786c0b241564936b4a9f0d6d52db8834064d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6998207
Commit-Queue: Javier Contreras <javierconmicrosoft.com>
Reviewed-by: Sam Davis Omekara <samomekarajrmicrosoft.com>
Reviewed-by: Alison Maher <almahermicrosoft.com>
Cr-Commit-Position: refs/heads/main{#1526543}

--

wpt-commits: dbbea4504a74c59e5050359d479e1e33b646f028
wpt-pr: 55287

UltraBlame original commit: efcd1898a20b0cf5b93330732764b0928ea0ea4c