Bug 1992486 - Add a probe to determine if the Recent Activity section is enabled. r=home-newtab-reviewers,nbarrett
There's a race condition with runtime ping and metric registration, where the TelemetryFeed
initialization script might execute (and attempt to use the highlightsEnabled metric) before
it has had a chance to be registered in Firefox version 144 and 145.

We work around this by adding a Promise that resolves when NewTabGleanUtils has completed
registration, which TelemetryFeed will wait for.

That workaround is fine for now, but the plan is to fix this more systemically for early-use
metrics and pings in bug 1992522.

Differential Revision: https://phabricator.services.mozilla.com/D267602

UltraBlame original commit: 5027de56a726479f96f141de564857b8a130d676