Bug 1994995 - Add the 'java' feature to the debug preset and make sure that every preset has it r=mleclair,android-reviewers,nalexander
This preset was added in bug 1992009, but it was missing the 'java'
feature. It's crucial to have this for two things:
1. We should always make sure to profile Android code when profiling on
   Android.
2. The profiler start logic actually doesn't work when this feature is
   missing.

The profiler start logic doesn't work because GeckoJavaSampler.isProfilerActive
method checks if the JVM sampler has started or not. We could call the
native profiler function using a JNI call but this brings an additional
overhead and this method needs to run very quickly.

Differential Revision: https://phabricator.services.mozilla.com/D269228

UltraBlame original commit: 12a7b6af5ad696a8a5424cb809c857d7dddcea0b