Bug 1995752. Make the async scrollable ancestor of fixed pos more consistent if there is no root scroll frame. r=botond
This is an edge case just to make sure things are handled in a consistent manner, it came up while working on anchor pos and making sure everything was sound.

There are still cases where we don't construct a root scroll frame, see https://searchfox.org/firefox-main/rev/a5316cedc669bcec09efae23521e0af6b9d3d257/layout/base/nsCSSFrameConstructor.cpp#2691 and the following code.

If we happen to have fixed pos in that case I think it makes more sense to continue looking for an async scrollable ancestor then returning null. For example we could have a parent doc and a scroll frame in the parent doc would then scroll this fixed pos. The parent frame of fixed pos is the viewport, so we'll go almost directly to the parent doc after this.

For example for how this would be more consistent, say we have a doc with no root scroll frame, and we have some fixed and nonfixed content in it, and a parent doc with a root scroll frame. Before this patch the asr of the fixed content would be null, the asr of the nonfixed content would be the root scrol frame of the parent doc. After this patch the asr of both would be the root scroll frame for the parent doc.

Making this consistent means that if two frames have the same scroll frame (of any type, not just async scrollable) then they will have the same async scrollable scrollframe as well. Which is the situation that I wanted to ensure true for anchor pos.

Differential Revision: https://phabricator.services.mozilla.com/D269578

UltraBlame original commit: e34970cedd0a4b58b224ec898de013f975eb12c2