Bug 1983388 - Prevent loss of focus when navigating to search pane. r=home-newtab-reviewers,akulyk,maxx,mconley
By changing the vbox id of the "Web Search" section, we prevent focus
from being stolen. I'm not entirely sure why this change fixes the issue
as it's likely something to do with the "7.4.6.4 Scrolling to a
fragment" part of the HTML spec. If I had to guess...it's because we end
up finding a valid focus target, due to the id of the vbox being "search".
Then we end up following the focusing steps and move the focus to the
document body.

I would bet there are no elements with ids that are the same as the hash
values we use to move between categories in about:preferences, thus
focus doesn't move because there are no valid focus targets in the
document.

See Comment #8 by :smaug for more detail.

Differential Revision: https://phabricator.services.mozilla.com/D265316

UltraBlame original commit: 5261daa46e9d087f616b74d0fcc5c128bb89f434