Bug 1995958 - Add a rudimentary texture viewer to the WR debugger. r=gw
This adds the 4 categories of textures to the documents panel with a button to fetch textures per category.
The textures can then be selected and previewed with a very rudimentary ui in the viewer panel.

It also adds the following commands: get-textures, get-atlas-textures, get-standalone-textures, get-target-textures and get-tile-textures.
When used from the cli, the textures are saved as pngs in the current folder.

The texture viewer ui is too basic to be really useful. I would like it to provide ways to zoom, color pick and save the textures at the very least, but that's a start.

Differential Revision: https://phabricator.services.mozilla.com/D269735

UltraBlame original commit: 978531406bc93346d64251b0531bdc1167681603