Bug 1996324 - Part 1: Move about:home startup cache tests under browser/components/newtab. r=home-newtab-reviewers,thecount
What's being tested here isn't really a function of the newtab extension, so much as the
function of the AboutHomeStartupCache and AboutNewTabRedirector modules.

The cache worker script still needs to live under browser/extensions/newtab in order to be
able to be updated in case one of the dependencies to render the cache changes via train-hopping,
so we'll keep the test_AboutHomeStartupCacheWorker.js test where it is.

Differential Revision: https://phabricator.services.mozilla.com/D270038

UltraBlame original commit: 30d8b45b19e77c42af03273f11e3cd2836c82504