Bug 1996182 - Fix up focus within state on focus redirect as well. r=smaug
This is... Not particularly pretty, to be clear. Maybe I should file a
spec issue to get better define when and how the :focus-within state
gets updated, because on one hand it's weird that when in a tree like:

<div id=parent>
  <input id=a>
  <input id=b>
</div>

Moving focus from a to b would allow you to observe parent without
:focus-within. But on the other the current set-up is clearly
error-prone.

Maybe we should add more state to nsFocusManager (an stack with
currently :focus-within ancestors) and fix it up "more properly" when
the focus actually moves (which could be on the blur event of a), but I
think I'd like to at least have some draft spec text for how does that
look.

I filed https://github.com/whatwg/html/issues/11835 to discuss this a
bit more thoroughly.

Differential Revision: https://phabricator.services.mozilla.com/D270166

UltraBlame original commit: 93b9cf7d4f69cd2322c7d9289508ba6b23665ae9