Bug 1919213 part 3 - avoid jumps and calls inside 13 bytes r=yjuglaret,win-reviewers,handyman
If there is a jump or call inside the first 13 bytes of the function that
we're detouring, we don't handle that case correctly. This should be very
unlikely, but now we will fail to detour. Also fixes up a test that was
attempting to do this.

Differential Revision: https://phabricator.services.mozilla.com/D268653

UltraBlame original commit: 3274320cf3c98e4e1d2d9e2743cbdd0bf3fca747