Bug 1995939: Don't clobber nargs before pushing undefined arguments r=jandem
Technically the change in emitAllocateSpaceForApply isn't necessary, because we'll end up storing `undefined` in the padding we've just allocated. However, loadFunctionArgCount is a load (guaranteed to hit in L1) and a shift, and we'd be doing a shift anyway. One L1 load is probably cheaper than an extra iteration of the undefined loop.

Differential Revision: https://phabricator.services.mozilla.com/D270058

UltraBlame original commit: 4af69c98454db9396d22bf9247f4fd31db90cb85