Bug 1996134 - Exclude aria actions targets from name calculation. r=Jamie
There is a chance of extra name change events in the case where the
aria-actions attribute is modified in a way that does not alter the
subtree targets.

To be fair, this is true with aria-labeledby as well when spurious
IDREFs are inserted into the attribute's string.

Differential Revision: https://phabricator.services.mozilla.com/D270199

UltraBlame original commit: d4911474b10df9a6c2814d9df7113cf4d36fbeed