Bug 1990634: Part 3 - Add WindowsTestDebug component to find open file handles r=win-reviewers,gstoll
The component is (only) useful for debugging when Windows tells us a process
has an open file handle and forbids operations on that file.  The utility
will return a list of applications with a handle to the file.  This was written
to debug this bug, but this bug was worse than it could handle -- file handles
can be cloned for memory mapping and it seems that the Restart Manager (the
Win32 tool we use to implement this) does not catch those file handles.

Process Explorer can do this as well.  (Pro tip: it even shows the memory
mapping handles... in the DLLs tab!).  This component is most useful for
fast, transient issues (such as Windows Defender grabbing a handle at an
inopporitune time), or for times when interactivity isn't available (such
as failures in automation, like this bug).

Differential Revision: https://phabricator.services.mozilla.com/D267710

UltraBlame original commit: b6938b1abadae16a947ea2345420495e5a68ff45