Bug 1994674 - Adjust sizes for flips. r=layout-anchor-positioning-reviewers,layout-reviewers,firefox-style-system-reviewers,dshin
We need to adjust for anchor-size(height) becoming anchor-size(width)
and so on. Luckily the infrastructure added on the previous patch
extends nicely to this.

And make the trait work with a mutable ref because it makes the calc
code a bit easier and more compact.

Differential Revision: https://phabricator.services.mozilla.com/D270595

UltraBlame original commit: 37dcda84a0f2865612d6742ddd166dd94c527280