Bug 1997118 - Mark gray bits as invalid if we abort GC during marking r=sfink
We depend on GC marking to propagate gray umarking through zones that are being
marked incrementally. Aborting this raises the possibility of creating black to
gray edges as previously happened with this testcase.

Marking the gray bits as invalid will cause the CC to perform another GC to
redo the marking. Aborting GC is pretty rare in practice so I don't expect this
to have a performance impact.

Differential Revision: https://phabricator.services.mozilla.com/D270546

UltraBlame original commit: d5d4c4541e46f7529840873812db128018f918d3