Bug 1993043 - Enable layout.css.text-decoration-inset.enabled by default r=layout-jp-market-reviewers,jfkthame
This also fixes the entry in property_database.js, which has a few problems.
The first value of `other_values` is the same as the initial value, the value
`calc(0)` does not parse as a length and `calc(0px) 0` will just become `0`
as that is the shorted serialization of that value, and according to the
WebVTT spec text-decoration-inset should not apply to ::cue.

Differential Revision: https://phabricator.services.mozilla.com/D270377

UltraBlame original commit: a086e60c24820c066d342589d5de30c0964c35c4