Bug 1996291 - Rework getStatus to getStatusByEngineId and fix status reporting in about:inference r=firefox-ai-ml-reviewers,tarek
When going through the tests I could not figure out what this code did.
After some sleuthing I realized that it was gathering information for
the about:inference. However, the code was badly broken, and incorrect
in what it was doing. I simplified the implementation a bit, and fixed
the breakage in about:inference.

There are no tests for this behavior in about:inference, and there as
far as I can tell there are no live models in about:inference that could
be used to test this behavior. Since it's a developer-only facing
feature, I'm not planning on adding a test for it.

Also please note that I'm adding an ml.d.ts file here to centralize some
shared types. See the in-flight TypeScript documentation here:
https://phabricator.services.mozilla.com/D268571

Differential Revision: https://phabricator.services.mozilla.com/D269982

UltraBlame original commit: 761e7a9fb4ca246c10fe70defa6d75f47109e135