Bug 1996291 - Adjust the flow for accessing the MLEngine to make the types correct r=firefox-ai-ml-reviewers,tarek
This changes the implementation a bit, but the flow should have the
exact same behavior. The ready status is not being applied in a promise
which should run before any other code. You can still check if the
engine is ready by awaiting on the engine. However, now when you need
the engine you just await the #engine property.

The documentation for the type was incorrect to what it actually was as
the #engine was original wrapped in a promise, then it was unwrapped,
which was a bit confusing to work with.

Differential Revision: https://phabricator.services.mozilla.com/D269981

UltraBlame original commit: db978f88a4f5eb884b40ddbd826f9418dc78d929