Bug 1997569 - When the decorating frame is an inline, use its content (not border) rect as the basis for text-decoration-inset computation. r=layout-jp-market-reviewers,layout-reviewers,AlaskanEmily
This makes sense because it is the content area within which the text
is positioned and has the decoration drawn.

Fixes the underline on the first few examples in the testcase (but not
the more complex ones; see following patch).

Differential Revision: https://phabricator.services.mozilla.com/D270842

UltraBlame original commit: 1d0ab889bdf27efc99dde7b445f8b3ebe402c33f