Bug 1996304 - Rename AUTO_PROFILER_TRACING_MARKER to AUTO_PROFILER_MARKER. r=mstange,profiler-reviewers,layout-reviewers,emilio
This also drops the category argument. The category string was confusing
because it ends up a field of the marker and not the actual category which is
an enum. Having the category as a field isn't really helpful.

This also sets us up for unifying AUTO_PROFILER_MARKER and
AUTO_PROFILER_MARKER_UNTYPED which become basically the same.

Differential Revision: https://phabricator.services.mozilla.com/D269992

UltraBlame original commit: 1e03cc1e2180aa3940d710bc673cd03ca9657f30