Bug 1996997 - Prefer the new deployment, use the fallback only if the new deployment doesn't answer. r=taskgraph-reviewers,ahal
This is the second step in the move to the new deployment.
The new deployment is consistently faster, so let's just use that by default.
To be on the safe side, we still want the old delpoyment as a fallback for a few days.

Differential Revision: https://phabricator.services.mozilla.com/D270881

UltraBlame original commit: 3fdcfdd1bd2d7a9ccd0d0ffe336b2dbd3f1b5084