Bug 1995637 - Make sure debugger object gets traced when tracing breakpoints r=iain
Breakpoints currently trace |wrappedDebugger| but this doesn't keep |debugger|
alive if we have nuked the CCWs. The debugger has a linked list of breakpoints that
each breakpoint is part of so we need to make sure it lives at least as long as
that.

The patch traces the debugger's object even if our CCW to it has been nuked.

Differential Revision: https://phabricator.services.mozilla.com/D271062

UltraBlame original commit: ca2f1c59f1ca9e38fac525496e142301b8009b41