Bug 1997389 - Post: Test that BackupService handles features being disabled at startup. r=cdupuis
This is meant to keep the strict checks on the number of observer
notifications while making it possible to check startup behaviour.

initStatusObservers and uninitStatusObservers are made usable in tests,
since otherwise creating a new BackupService would add more
notifications every time something changes. The design is inspired by
initBackupScheduler and is called automatically for the global instance;
as such, no UI code needs to change.

The spurious infrastructure is provided for the next commit in the
stack, although it makes me wonder whether the strict counting is really
useful. (In this patch it also accounts for a Nimbus design choice.)

The removed tests are obsolete thanks to those in
test_BackupService_enabled.js.

Differential Revision: https://phabricator.services.mozilla.com/D270899

UltraBlame original commit: cad1205782aa3785a86e227055f39b37015534ac