Bug 1635774 part 3: Add about:keyboard UI to customise keyboard shortcuts. r=fluent-reviewers,mossop,bolsson,toolkit-telemetry-reviewers,frontend-codestyle-reviewers
This is an unpolished UI for experimentation.
It is likely that this UI will be heavily modified or even completely replaced before it is considered official.
As such, this is only available via a new about:keyboard page with no entry point elsewhere in the UI.

This primarily obtains keyboard shortcuts from the menu bar menus.
This doesn't allow all keyboard shortcuts to be customised, since many aren't available in the menu bar menus and haven't been hard-coded here.
There is a basic framework to allow for hard-coding of additional shortcuts, though, and this has been used for a few shortcuts already.

For now, when a key combination involves a character and a modifier changes the character, the UI shows the modified character.
For example, on a US English QWERTY keyboard, shift+3 produces the `#` character, so the UI shows `Shift+#`.
While this is likely to be a little confusing for some users, this is difficult to fix in a locale independent way.
The current behaviour should allow this to work across locales, albeit with this somewhat quirky presentation.

Differential Revision: https://phabricator.services.mozilla.com/D266792

UltraBlame original commit: 6d5ca7a5139782ef8cab7ff06dea85669f93b567