Bug 1968521 - Add subtest alerting for jetstream3 subtests with alert threshold at 5% r=perftest-reviewers,sparky,jandem
This patch enables alerting on JetStream 3 benchmark. For subtests, 5%
threshold will be used, and the "overall score" will still be 2%. As JS3
is still in active development, 2% could be too noisy for the subtests
so for the time being 5% will be used.

Additionally, a new "subtest_alert_threshold" attribute is introduced as
a better way to seperate alerting from main score and subtests.

Differential Revision: https://phabricator.services.mozilla.com/D271141

UltraBlame original commit: 53edeae7908a1c1f4bcca779ab72470e90c2ee2e