Bug 1996245 [wpt PR 55632] - [Navigation API] navigate-replace-same-document.html is written in a way that doesn't catch exceptions, a=testonly
Automatic update from web-platform-tests
[Navigation API] navigate-replace-same-document.html is written in a way that doesn't catch exceptions (#55632)

This test is written in a way such that the harness does not catch the exception.
Currently, WebKit hits one of the asserts in the test, but the test still passes
because the harness isn't catching the exception. We rewrite the test in a way
that allows the exception to be caught and the test to fail properly.
--

wpt-commits: 65f613b5da6a1bbf8a58efc35951f0a89e6589a7
wpt-pr: 55632

UltraBlame original commit: 3d103d24d48fb88868695b42258dcf3efcc6bb1c