Bug 1998178 - Switch default 'run-on-repo-type' to both Git and Hg, r=taskgraph-reviewers,mozperftest-reviewers,releng-reviewers,jcristau,sparky
The taskgraph diff is identical as we add an explicit
`run-on-repo-type: ["hg"]` to every task. This way it's easier to grep
for which kinds still haven't been migrated yet, and new kinds will be
more likely to work with Git from the get go.

Differential Revision: https://phabricator.services.mozilla.com/D271271

UltraBlame original commit: 0373031bfeac169bbc9e14b83e2fb340f811280b