Bug 1982378 part 2 - Fix Debugger.Script.getOffset{Metadata,Location} return value for prologue bytecode ops. r=arai
If the offset was in the prologue, `getOffsetMetadata` and `getOffsetLocation`
were returning information for the first 'main' op instead. This could result in
confusing stepping behavior in DevTools if the first main op is a valid stepping
location, because the debugger would pause in the prologue.

This patch adds a `SkipPrologueOps` argument to `BytecodeRangeWithPosition` and requires
callers to specify whether they want to skip prologue ops.

Differential Revision: https://phabricator.services.mozilla.com/D271579

UltraBlame original commit: 80e3deb7171ca4e6563159ba15b51497b34bd18a