Bug 1998993 - [ci] Ensure 'fetch-content' script is mounted for tasks with use-uv / use-python, r=taskgraph-reviewers,bhearsum
There's a check that only mounts the fetch-content script if a task is
using fetches:
https://searchfox.org/firefox-main/source/taskcluster/gecko_taskgraph/transforms/job/run_task.py#195

However, the use-uv and use-python flags will cause fetches to get added
later on. So far this hasn't broken anything because run-task doesn't
actually use the mounted version of fetch-content anyway, it hardcodes a
path in the checkout:
https://searchfox.org/firefox-main/source/taskcluster/scripts/run-task#581

But now we're starting to run tasks that are using the upstream version
of run-task, and that does require fetch-content to be mounted in the
same directory as run-task.

Differential Revision: https://phabricator.services.mozilla.com/D271820

UltraBlame original commit: c7b59fa11a60115447d71a58d21f60f513ea5b1b