Bug 1986285 - Part 6: Reduce dependency on fixed frecency values in test - r=daisuke,home-newtab-reviewers,nbarrett
Since the value of frecency in the table is the days until the score for a website
becomes 1, the values are a moving target. Thus, we should look to do basic
comparisons in the test.

Additionally, the ordering of the reset was causing TV errors.

Differential Revision: https://phabricator.services.mozilla.com/D264740

UltraBlame original commit: 4e30c900c891d32be2cb23385ee612b0f518d5ff