Bug 1998716 - [remote] Fix jsdoc/valid-types issues in remote r=frontend-codestyle-reviewers,Sasha
When an event name contains special characters, which can't normally be used in JS variable name (eg `-`, `.`), the event name must be escaped with surrounding quotes. However when doing this, the namespace (normally optional for `fires`), becomes mandatory.
So the pattern to use is `fires ClassName#"event-name.with-special.chars"`. But when using a simpler name (eg `message` or `beforeStopRequest`) it can be directly documented without namespace or escaping (`fires message`).

We can align the other fires for consistency, but I would file something else to do so.

Differential Revision: https://phabricator.services.mozilla.com/D272081

UltraBlame original commit: 1efdc3046108b08a8271a5de1406d141f5526aef