Bug 1827612 - Remove stable hashing for WeakMap and rekey entries instead r=sfink
This adds a vector of keys of entries where the key or value is in the nursery
so that we can trace these at the start or minor GC. If the vector grows too
large we sweep the whole map instead.

This replaces the post barrier for WeakMap keys so the table now uses
PreBarriered<Key> instead of HeapPtr<Key>.

This doesn't provide much improvement to WeakMap.get performance on its own but
it is a prerequisite for bug 1995021.

Stable hashing is still used for WeakRef and FinalizationRegistry targets.

Differential Revision: https://phabricator.services.mozilla.com/D270562

UltraBlame original commit: 7cc46dd9bdfaafeaf9905a575697366b4f3f07a8