Bug 2000025 - commit sha guidance should be in oldest-to-newest order for git like hg. r=dbaker DONTBUILD
It makes it far more tedious if using that last of commit shas as
input to extract-for-git.py if they're in the reverse order.  When
running under mercurial, the order is already correct for using with
extract-for-git.py.

Differential Revision: https://phabricator.services.mozilla.com/D272512

UltraBlame original commit: 0f85a0ea3f27c07ff56a53c97b8877a71f1ea020