Bug 1993474 - remove WS_BORDER to fix border issues with Windows App SDK r=win-reviewers,handyman
When a window with WS_BORDER is maximized on a low-ish DPI monitor, an
"extra" border on top shows up. This breaks the workflow of moving your
mouse to the very top of the screen to hover over or select tabs.

(people have reported this with non-Firefox apps so I'm guessing this is
really an OS problem)

But the WS_BORDER style doesn't actually seem necessary for us when tabs
are being shown in the toolbar - removing it still means we can resize
our window, etc.

We do need to handle the case when tabs are not being shown in the toolbar,
so we have to toggle the WS_BORDER style when that changes. I tested on my
system in a number of different configurations and it seems to work fine.

I believe this will fix bug 1994918 and possibly some others.

Differential Revision: https://phabricator.services.mozilla.com/D272653

UltraBlame original commit: 6b6fbc2c8e4cfc2433d377f99c134190527bf608