Bug 2000496. Cleanup SpecialPowersParent.flushPrefEnv and popPrefEnv. r=emilio
popPrefEnv doesn't do anything async, and we are not awaiting it either, so don't make it async.

This meant that |this.popPrefEnv().requiresRefresh| is a bug in that is tries to get requiresRefresh on the promise that popPrefEnv was returning (which didn't exist).

And then make sure to turn requiresRefresh back into a boolean instead of a number. I don't think this caused any problems but for consistency.

Differential Revision: https://phabricator.services.mozilla.com/D272829

UltraBlame original commit: 10a635d9c5d3cb132300e1b43fc8ce1ee2f3f7ac