Bug 1999986 - Remove ToSecondsSigDigits from TimeDuration. r=glandium
ToSecondsSigDigits was always intransparent about the actual precision wrt ToSeconds and not really helpful in deciding, what to show to a human reader.
Now that both on Windows and MacOS the precision is always the very same as that of ToSeconds and after switching the last consumers to use just that, we can just remove it alltogether.

Differential Revision: https://phabricator.services.mozilla.com/D272842

UltraBlame original commit: 5e82dc08799bb2be710d7cb8d4e589d349bf62be