Bug 1989604 - Fix sync button loading forever on Tabs Tray r=android-reviewers,pollymce
WorkManagerSyncManager's workersLiveData might skip the RUNNING emission if WorkManagerSyncWorker.doWork's execution is really quick (in the case where the last sync was within the past 5 seconds and the sync operation is skipped). The previous code was relying on RUNNING to be received to start set isSyncActive to true, and would only dispatch onIdle() when receiving a non-RUNNING state if isSyncActive was already true.

The new implementation should be more reliable, as it will trigger onIdle when receiving a "finished" state, even if the running state wasn't previously received. With that, we make sure we correctly stop the loading state on screen.

Differential Revision: https://phabricator.services.mozilla.com/D272098

UltraBlame original commit: c4ebaa7fb78188899b62255cfb02ef1bfe3f04b8