Bug 1999784 - fix crash on add password. r=android-reviewers,mcarare
This is very similar to the previously fixed scenario on edit password - a store action was being dispatched using io scope, then it does some ui stuff that requires the main thread. Thread management is no longer handled as a side effect of the Store type you choose, so we need to be explicit here and launch the ui code on main.

Differential Revision: https://phabricator.services.mozilla.com/D273418

UltraBlame original commit: 207e80fc53726ca60ec8406002f9814bec3b56a0