Bug 2001562 - [devtools] Fix AnimationPlayerActor for element with `type` property. r=devtools-reviewers,bomsy.
In the class, there were a couple places where we would check the animation target
`type` property as a way to detect if the animated element was a pseudo element.
This means that we were considering button elements as pseudo element, and re-assigning
variables with erroneous properties, which would ultimately make the methods throw.
This is also wrong for pseudo elements: the animation target is always the binding
element and we have a `pseudoElement` string property on the animation to detect
if this applies to a pseudo element.

We fix the faulty callsites and add a server test that is covering this.

Differential Revision: https://phabricator.services.mozilla.com/D273546

UltraBlame original commit: 2d15c1d5edd15aa813c09ef16f6b573c7c90940c