Bug 1999841. Check for failure getting data from DataSourceSurface in GetSkImageForSurface. r=lsalzman
The surface is a SourceSurfaceWebgl. When we call SourceSurfaceWebgl::EnsureData the only thing present in the surface is a mDT weak pointer. The pointer is evidently gone as when we try to use it in the pernosco we are not able to. So the data and stride are 0.

Differential Revision: https://phabricator.services.mozilla.com/D273387

UltraBlame original commit: e1be9bd1c19edfefd57bb95b29b7e2ddd72c363b