Bug 2001395 - Set Fx Backup prefs to enabled for Windows by default r=nrishel
Instead of defaulting the availability of Fx Backup to false, we
default it to true on Windows. There is still a Nimbus setting
in place that controls whether or not Fx Backup is actually
shown to the user.

Differential Revision: https://phabricator.services.mozilla.com/D273927

UltraBlame original commit: 8a90d29d9973c57cf516151d0ebe5cfcbd164dd7