Bug 2001721 - Prevent busy loop when tunneled connection returns WOULD_BLOCK, r=necko-reviewers,valentin
When a tunneled connection returns WOULD_BLOCK, the consumer calls AyncWait() and we enqueue the stream on the ready queue. We then call ForceSend() -> ProcessOutput() and immediately attempt to send again, which still would block, causing the stream to be re-queued and retried in a busy loop.

This patch introduces a mBlockedByFlowControl flag that is only cleared when neqo notifies us the stream is writable. While the flag is set, we skip send attempts. This avoids the busy waiting.

Differential Revision: https://phabricator.services.mozilla.com/D274000

UltraBlame original commit: 99dc6355eb923c211d758616ad22623d9eb0e2b9