Bug 2002825 - Add a test for the sourceId of JIT frames inside marker stacks r=mstange,profiler-reviewers
This test checks if we have sourceIds in the JIT frame of testFunction
when they are captured from markers. Previously, we were incorrectly
not collecting the sourceId if the samples are collected for the marker
stacks. Now, we properly collect them.

Differential Revision: https://phabricator.services.mozilla.com/D274330

UltraBlame original commit: c64c385d13acdf2eebeb2ff1aacd76078dabed20