Bug 1999559 - Specify `interactive-widget=resizes-content` in helper_zoomToFocusedInput_scroll.html. r=botond
This test was added in bug 1149300 when we set
`ScrollFlags::OverflowHidden` for zoom-to-focused-input.

At the time the test was added we had no interactive-widget support,
our default behavior was `resizes-content`. Thus the test should work
with `interactive-widget=resizes-content` as expected what the test was
supposed to be at that time.

Differential Revision: https://phabricator.services.mozilla.com/D272183

UltraBlame original commit: bcc17dbf5d94a6382a7dea33303eab528c2410b5