Bug 2003127 - Make the layout module paint normal selection when selection is styled as "attention" r=emilio,layout-reviewers
The "Find in Page" result appears as a normal selection range and the
selection's display is "attention".  And that may select characters in
unselectable elements including `::before`/`::after` pseudo elements.

At least while the selection's display is "attention", we should paint
the normal selection in any nodes.

Note that once the selection's display is set to "on", the selection
becomes invisible since the user not selectable and won't appear as
visible text when you copy the selection to the clipboard. If this is
a bad UI, we should change the behavior in a follow up bug.

Differential Revision: https://phabricator.services.mozilla.com/D274550

UltraBlame original commit: 9336a9aece477c6a071a5f90693f72f0cb32fcf0