Bug 1993407 - Use the visual viewport size not including the dynamic toolbar for a specific case. r=botond,geckoview-reviewers
We use the visual viewport size including the dynamic
toolbar area in APZ generally, but it has a side-effect happening in
FrameMetrics::KeepLayoutViewportEnclosingVisualViewport. In
KeepLayoutViewportEnclosingVisualViewport we try to adjust the layout
scroll offset so that the updated visual viewport by the new visual
offset fits within the layout viewport range. The side-effect is
harmless in most cases but we recently realized that it cases a user
visible issue such as bug 1997575.

To avoid bug 1997575 specifically, this change uses the visual viewport
not including the dynamic toolbar in KeepLayoutViewportEnclosingVisualViewport
for the specific case where all following conditions are met;

i) the dynamic toolbar is fully visible
ii) the software keyboard is opened
iii) the interactive-widget mode is resizes-visual

Differential Revision: https://phabricator.services.mozilla.com/D272985

UltraBlame original commit: 781114d3fb0b22584c555af1c8aeeb0fcbe14506