Bug 2003588 - Continue to allow creation of CCWs to debugger instances after CCWs have been nuked r=jandem
It turns out that disallowing this breaks the checks in Compartment::wrap. I
don't fully understand what the implications of this are but it seems fine to
continue to allow creating these if we don't nuke them in the first place.

Differential Revision: https://phabricator.services.mozilla.com/D275046

UltraBlame original commit: 0dba07ef81b568b78fcc398303d8d94e584ec1d1