Bug 1987375 - Remove text/plain flavor when dragging a download to work around Adobe Illustrator r=handyman
When the text/plain flavor is included in a DataTransfer object that is
dropped over Adobe Illustrator, Illustrator prefers that to the file flavor,
and so a text version of the URL is dropped.

This is arguably a bug in Illustrator, but when drag/dropping from Chrome
they don't include the text/plain flavor, so it seems fine to omit it.

Differential Revision: https://phabricator.services.mozilla.com/D275066

UltraBlame original commit: a194000cb5119b9d725575595dffd96ed46ea32a