Bug 1873716: Handle about:blank iframes that do fingerprinting r=timhuang
We encountered a fingerprinter that created an about:blank iframe
and did fingerprinting there. This iframe had no channel, so it
did not get a ContentBlockingEvent on it.  Traverse the document
tree upwards to find a document we can log it on.

SKIP_BMO_CHECK

Differential Revision: https://phabricator.services.mozilla.com/D274380

UltraBlame original commit: 33eeafd6cdf272db573df723f95e027f1e25fb21