Bug 2004126: Maintain min-size for CCGraph::mPtrInfoMap r=mccr8
The approved patch in bug 1477627 was not the one that landed.  The landed
code fails to maintain the min-size of an internal GC pointer map (16K)
and, after the first GC, reduces it to the min (4... as in 4, not 4K).  This
causes the GC to constantly do rehashes, which constitute over 20% of time
taken by GC in a profile with a very long (10 second) GC.

The patch in bug 1477627 also used HashSet::initialized in CCGraph::IsEmpty,
but that seems wrong and no longer exists anyway.

Differential Revision: https://phabricator.services.mozilla.com/D275109

UltraBlame original commit: c8a13116b30c04921d2ce9d9a71950db5162ddc0