Bug 1998182 - Add a way to list draw calls and other interesting renderer commands in the debugger. r=gw
This adds a new Draw calls panel in the debugger and some infrastructure for adding per-frame logging to be displayed in the debugger UI.
This will be followed up with a timeline widget to chose which frame's log to display.

The goal is to extend this to log and display any information for which it is useful to see changes over time or for which capturing the right frame is tricky.

Differential Revision: https://phabricator.services.mozilla.com/D271238

UltraBlame original commit: c1865ed5009d4a8415511ce311d844431a9aca12