Bug 1892201 - Add support for caching quad render tasks. r=gw
And use it for conic gradients with SWGL.
This takes a simplified approach for the cache key, only considering the item UID and up to 3 clips.
As a result the cache key is small and independent from the primitive kind. The downside is that it cannot deduplicate identical items within a frame.

Differential Revision: https://phabricator.services.mozilla.com/D275218

UltraBlame original commit: 0080d2ecc137871dd3d349d2e3239240905819ab