Bug 1994828 - Track wasm mime type essence in ModuleLoadRequest; r=allstarschh
The only way we can distinguish between JavaScript and wasm modules is
using the mime type in the http response. This adds a field to track this
information in the load request, so we know later on if we should
attempt to compile the module as JavaScript or wasm.

Differential Revision: https://phabricator.services.mozilla.com/D270150

UltraBlame original commit: 467c90d79ed5546af8c7c6b49c4961148d7d6346