Bug 1817051 - Part 3: Start to sample scroll timeline in the HTML loop. r=firefox-style-system-reviewers,layout-reviewers,hiro,emilio
Also, we drop ProgressTimelineScheduler since we don't need it anymore.

The update in the test is necessary (just like what we do for wpt) for
make sure getComputedStyle() works after we sample the scroll timelines,
because the updating of stale timelines has no effect on forced style and
layout calculations triggered by getComputedStyle() or similar.
In other words, initially stale timelines are visible as such through those
APIs.

Note: there are two new test failures because the timing of caching the
current time is later than the moment we trigger the animation and
scrollTo API. I will try to fix them in the following patch. In this
patch we focus on switching the sampling way and to make sure we don't
cause lots of regressions.

Differential Revision: https://phabricator.services.mozilla.com/D269666

UltraBlame original commit: 1cbc83b410ab9705ff63e050df1d90196b6ed4cf