Bug 1927599 - Part 6: Get OpenGL-based rendering working on-device, r=gfx-reviewers,ahale
This requires adding a call to texImageIOSurface, which is only available
on-device, so cannot be used for simulator builds. We will need to come up with
another soltuion for simulator, which may reqire a non-IOSurface-based backend.

In addition, the compositor code needs to be changed to use GL_TEXTURE_2D
instead of GL_TEXTURE_RECTANGLE, as that type doesn't exist for OpenGL ES on
iOS.

Differential Revision: https://phabricator.services.mozilla.com/D229403

UltraBlame original commit: e562cebaefcd90cf44a04877b14ce2edde49fb64