Bug 1817051 - Part 1: Add AnimationTimelinesController. r=layout-reviewers,hiro,dom-core,smaug
We wrap the list of timelines attached to the document into a controller,
so this makes it easier to support diffrerent types of timelines, e.g.
scroll-timeline.

We will add the implementation of scroll timelines in the following
patches.

Differential Revision: https://phabricator.services.mozilla.com/D269663

UltraBlame original commit: 4a800fe06f574e523aa929f5cb3e56835369ca53