Bug 2001722 - Simplify X11 client margin computation. r=stransky
If there's no client area we're using topLevelBounds, but that's not
right since we expect the bounds to be relative to topLevelBounds (i.e.,
the bounds should've been moved to 0, 0.).

Fix the issue, but also derive the client margin from the client area
rather than the other way around, to avoid this mistake causing just
missing content.

This matches what we do for wayland and in general should be saner.

Differential Revision: https://phabricator.services.mozilla.com/D276552

UltraBlame original commit: d5771455b703603414239609b136ca4458549f72