Bug 1989246 - Make font-unicode-PUA.html more robust. r=jfkthame
font-unicode-PUA.html renders PUA characters with a font-family list
made of generic families except the last one. It verifies that this
perfectly matches the same text rendered with only this last non-generic
family, based on this requirement from the CSS spec [1]:

"If a given character is a Private-Use Area Unicode codepoint, user
agents must only match font families named in the font-family list that
are not generic families"

However, nothing in the spec actually prevents the "first available
font" to be one of the generic font family, which can affects other
metrics (e.g. line-height). In particular, the test currently uses
'Times' as the last (non-generic) family, which likely does not provide
the requested PUA glyphs and so be rendered with a missing-glyph box
that again uses the metrics from the "first available font".

The test covers the generic font-family "math", and for the reasons
described in the previous paragraph, it regressed after [2].

Instead, we rewrite this as a testharness test to use the Ahem web font,
the three PUA characters it contains (U+F000, U+F001, U+F002) and only
check the glyph's widths, making the whole setup more robust.

[1] https://drafts.csswg.org/css-fonts-4/#char-handling-issues
[2] https://phabricator.services.mozilla.com/D256978

Differential Revision: https://phabricator.services.mozilla.com/D276810

UltraBlame original commit: b953a9fcd9a62597646afb4dae329133122239a6