Bug 2006613 - Plumb batch removal state through unbind, and use it to optimize Shadow DOM removal. r=smaug
This shows up in bug 2000414, where the front-end clears all the
menuitems of a <menupopup> at once, and they are all slotted, and
clearing the assigned node list from the front is more expensive than
necessary. See the profile in: https://share.firefox.dev/493zcYj

Of course, it'd be better to remove back to front, but that would be
observable, and it seems we can improve our batch removal machinery to
deal with this case quite easily, since if all the host kids are going
away, we know we will end up with no assigned nodes at all.

Differential Revision: https://phabricator.services.mozilla.com/D276828

UltraBlame original commit: 96766b6ad8025ea1af3e9bf817e598819bb8b984