Bug 2002127 - Part 17: Update Ghost Button Usage r=translations-reviewers,hjones
This commit updates the usage of ghost icon buttons in the Translations Settings
subpage to only be used in the case of showing a loading spinner, a warning icon,
or an error icon. If the button is intended to be clicked, whether enabled or disabled,
it will be shown as a non-ghost button.

Differential Revision: https://phabricator.services.mozilla.com/D274829

UltraBlame original commit: e6754b283b6615d2a2f714be848e1334ec397a59