Bug 2005582 - [devtools] Fix copying from markup view HTML editor. r=devtools-reviewers,ochameau.
With the migration to CodeMirror 6, the copy event listener callback that
is set in the Markup view would not see the element as being an editor (it
used to have a textarea for the CodeMirror editor, but not anymore), and
so proceeds to copy the outerHTML of the selected node instead.

This is fixed by checking if the event target is in a CodeMirror instance,
on top of checking if it's an input or a textarea.

A test case is added to cover the fix.

Differential Revision: https://phabricator.services.mozilla.com/D277026

UltraBlame original commit: 81c47bd3efa5421a0041c21dab2c1468f333ed99