Bug 2007058 Part 1 - Generalize NS_FRAME_IS_PUSHED_FLOAT bit to NS_FRAME_IS_PUSHED_OUT_OF_FLOW. r=layout-reviewers,emilio
We need a similar concept to mark an abspos frame that is pushed to a later
absolute-containing-block continuation. A frame is either a float or an abspos
and thus lives in a different frame child list, so the same bit can be used in
both scenarios without conflicts.

Differential Revision: https://phabricator.services.mozilla.com/D277137

UltraBlame original commit: a83d625717c8c1ec81cc1799308cd1126ecf7360