Bug 2006508 - [devtools] Show ::backdrop node in Inspector. r=devtools-reviewers,ochameau.
Accept ::backdrop pseudo element in the walker standard filter and set the proper
parent for it in getBindingElementAndPseudo.

We also add a few test cases to make sure the node is properly displayed in the
markup view and the animations panel.

Differential Revision: https://phabricator.services.mozilla.com/D276854

UltraBlame original commit: 40eb0448656a97059bbc44c90ff5cb5925ea8130